### 2025-05-14 #8201 @editorialbot
[PRE REVIEW]: Tesseract Core: Autodiff-native, self-documenting software components for Simulation Intelligence
**Submitting author:** @boltzmannbrain (<a href="http://orcid.org/0000-0003-3422-7820">Alexander Lavin</a>)
**Repository:** https://github.com/pasteurlabs/tesseract-core
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.9.0
