### 2025-05-14 #8201 @danielskatz
👋 @BoltzmannBrain - thanks for your submission.

However, I don't think I really understand what the software here is, what it is used for, or who would use it. Is there something in the documentation or repo that walks through some type of a simple research use case?  The paper currently is a little too abstract for me, and adding something more specific point to the documentation would be helpful. I suggest this because the paper is already too long, and adding more would not be a good fit for JOSS. In fact, if there's anything in the paper that can be removed in favor of pointing to the documentation or the repo, please do that as well.

Additionally, and as a minor point in comparison, please see the JOSS [example paper](https://joss.readthedocs.io/en/latest/example_paper.html) for how references that are directly addressed should be mentioned in the .md file.

Please feel free to make changes to your .md file, then use the command `@editorialbot generate pdf` to make a new PDF. editorialbot commands need to be the first entry in a new comment.
