+++ docs/index.md
@@ -102,7 +102,6 @@ content/using-tesseracts/advanced.md
 Expert showcase <https://si-tesseract.discourse.group/c/showcase/expert-showcase/12>
 How-to Guides <https://si-tesseract.discourse.group/c/support/how-to-guides/10>
 content/examples/example_gallery
-content/tutorial/bootcamp
 ```
 
 ```{toctree}
