+++ docs/content/tutorial/bootcamp.md
@@ -1,140 +1,0 @@
-Tesseract Bootcamp
-====================
-
-````{note}
-Complete this bootcamp to create and run your first very own Tesseract!
-````
-
-We at the Tesseract team have recently received a mysteriously encoded
-message hidden in our lounge room.
-
-:::{dropdown} Secret Message
-```{literalinclude} ../../../tests/tutorial/example_data/secret_message.json
-:language: json
-```
-:::
-
-
-Help us figure out what this ciphertext means -- in a Tesseract way.
-
-## Step 0: Initializing Your First Tesseract
-
-Create your own Tesseract named `bootcamp` by running the command `tesseract init --name bootcamp --target-dir bootcamp`.
-
-This will create a brand new Tesseract directory `bootcamp` with an empty api, config, and requirements file.
-
-Run `tesseract build bootcamp` to validate everything is working.
-
-:::{dropdown} Unit Test
-:color: secondary
-:icon: check-circle
-
-Run the test for the first step.
-
-* Make sure you run the build command first.
-
-`pytest tests/tutorial/test_bootcamp.py --tesseract-dir [path to bootcamp directory] -k test_00_tesseract_init`
-
-Example:
-`pytest tests/tutorial/test_bootcamp.py --tesseract-dir bootcamp -k test_00_tesseract_init`
-
-Make sure the first test is passing.
-:::
-
-## Step 1: Filling out the api
-
-````{note}
-
-Learn how to define your own input/output schema and write your first apply function.
-
-````
-
-Now, in your generated tesseract, there should be a semi-populated `tesseract_api.py` file.
-
-Populate it by following the instructions here: [Get Started - tesseract_api.py](#getting-started).
-
-For our `bootcamp` Tesseract, we want to be able to decode the ciphertext that is a numpy array of random integers into a hidden string message.
-
-**Populate the `input schema` and `output schema` fields** so we can achieve this, then make sure we pass the next test. The available schema types are defined in [tesseract-runtime-api](#tr-runtime-api), and you can also reference the examples in the [example-gallery](#example-gallery) for some inspiration. Also take note of how secret message is structured when defining your input types.
-
-Make sure you rebuild the bootcamp Tesseract before proceeding to the testing phase.
-
-:::{dropdown} Unit Test
-:color: secondary
-:icon: check-circle
-
-`pytest tests/tutorial/test_bootcamp.py --tesseract-dir [path to bootcamp directory] -k test_01a_tesseract_schema`
-
-example:
-`pytest tests/tutorial/test_bootcamp.py --tesseract-dir bootcamp -k test_01a_tesseract_schema`
-
-::::
-
-Now that the schemas are populated, we need to fill out the apply function. We believe that the ciphertext is encoded using a basic *letter number cipher (A1Z26)*. **Write the apply function so that we can pass in our ciphertext and get the decoded message back.**
-
-```{tip}
-Use examples/vectoradd and examples/helloworld as inspiration.
-```
-
-Test using the next unit test to validate we can decode some basic messages.
-
-
-:::{dropdown} Unit Test
-:color: secondary
-:icon: check-circle
-
-`pytest tests/tutorial/test_bootcamp.py --tesseract-dir [path to bootcamp directory] -k test_01b_tesseract_apply`
-
-You can also test this locally by running
-```
-tesseract run bootcamp apply @tests/tutorial/example_data/test_apply.json
-```
-::::
-
-## Step 2: Passing in the key
-
-````{note}
-
-Learn how to pass in and process local data in your Tesseract.
-
-````
-Now, let's try passing in our secret message into the Tesseract to see if it can be correctly decoded.. Copy and paste the secret message above
-into a file named `secret_message.json` and pass it into your Tesseract's
-apply function!
-
[…]

[…]
