### 2025-05-14 8393c63
add sha256 prefix if missing
