### 2025-05-14 #8201 @BoltzmannBrain
Hi @danielskatz, thank you for the speedy follow-up and helpful comments!

Yes we have tons of research use-cases:
- ["Showcases" in the Tesseract community forum](https://si-tesseract.discourse.group/c/showcase), for example this real-world plasma physics application of the software: [Thomson scattering diagnostic parameter optimization](https://si-tesseract.discourse.group/t/thomson-scattering-diagnostic-parameter-optimization/45)
- User docs that explain how and why to use Tesseract's automatic differentiation algorithms and operations, with emphasis on scientific machine learning and differentiable physics: [Differentiable physics programming in Tesseract Core docs](https://docs.pasteurlabs.ai/projects/tesseract-core/latest/content/introduction/differentiable-programming.html)
- and of course more getting added to the main repo and community forum every week.

We will address your other comments and push the updates ASAP this week.
