+++ joss-paper/paper.md
@@ -27,35 +27,43 @@ bibliography: paper.bib
 
 # Summary
 
-Tesseracts are universal software components for scientific computing, simulation, and machine learning (ML), summarized as “simulation intelligence” (SI) [@lavin-si]. Specifically, Tesseracts enable and expedite the transition from experimental, research-grade software to production environments. This includes native support for automatic differentiation between heterogeneous software artifacts in distributed and cloud contexts, which enables end-to-end differentiable programming, hybrid ML and simulation systems, and more for SI at scale.
+Tesseracts are universal software components for scientific computing, simulation, and machine learning (ML), summarized as "simulation intelligence" (SI) [@lavin-si]. Specifically, Tesseracts enable and expedite the transition from experimental, research-grade software to production environments. This includes native support for automatic differentiation between heterogeneous software artifacts in distributed and cloud contexts, which enables end-to-end differentiable programming, hybrid ML and simulation systems, and more for SI at scale.
 
-Tesseract Core is a Python library that serves as the reference implementation for defining, containerizing, executing, and deploying Tesseract components. It provides user entrypoints to wrap existing software artifacts such as Python functions, Julia code,  C++ libraries, or remote services into Tesseract components. By unambiguously defining allowed inputs and outputs of each Tesseract via Pydantic models [@pydantic], Tesseract Core enables external data validation and auto-generation of machine-readable API schemas. This allows users to explore the capabilities of Tesseract components without interacting with code, including their inputs, outputs, and available operations. Similarly, it allows workflow engines to compose them into larger, self-validating pipelines that are, by the virtue of implementing Tesseracts, end-to-end differentiable. Tesseract components are built to by deployed in distributed contexts, including support for containerization and remote procedure calls (RPC) via a REST API.
+Tesseract Core is a Python library that serves as the reference implementation for defining, containerizing, executing, and deploying Tesseract components. It provides user entrypoints to wrap existing software artifacts such as Python functions, Julia code, C++ libraries, or remote services into Tesseract components. By unambiguously defining allowed inputs and outputs of each Tesseract via Pydantic models [@pydantic], Tesseract Core enables external data validation and auto-generation of machine-readable API schemas. This allows users to explore the capabilities of Tesseract components without interacting with code, including their inputs, outputs, and available operations. Similarly, it allows workflow engines to compose them into larger, self-validating pipelines that are, by the virtue of implementing Tesseracts, end-to-end differentiable. Tesseract components are built to by deployed in distributed contexts, including support for containerization and remote procedure calls (RPC) via a REST API.
 
 # Statement of need
 
 The last few years have seen a rapid increase in the use of machine learning and data-driven methods within scientific computing and industrial simulation. However, when moving beyond the research lab and into real-world engineering workflows, the gap between (GPU-accelerated, autodiff-native, high-level, bare-bones) research code and (CPU-bound, non-differentiable, low-level, feature-rich) production code is often too large to bridge. Similar issues arise when integrating components of different software stacks (e.g. Python, Julia, C++), and when deploying components in heterogeneous or distributed environments (e.g. cloud vs. on-prem, HPC, and supercomputers).
 
-These software bottlenecks translate to significant untapped potential of differentiable programming in particular and simulation intelligence in general—for example, when exploring the merits of hybrid AI-simulator systems (e.g. in differentiable physics contexts such as [@neuralgcm;@solverintheloop;@diffphys]), or when using machine learning to augment existing simulation codes (e.g. when learning closure models as in [@closure1;@closure2]).
[…]
