### 2025-05-15 #8201 @danielskatz
> Hi [@danielskatz](https://github.com/danielskatz), thank you for the speedy follow-up and helpful comments!
>
> Yes we have tons of research use-cases:
>
> * ["Showcases" in the Tesseract community forum](https://si-tesseract.discourse.group/c/showcase), for example this real-world plasma physics application of the software: [Thomson scattering diagnostic parameter optimization](https://si-tesseract.discourse.group/t/thomson-scattering-diagnostic-parameter-optimization/45)

This is useful. Could the paper text point to this?

> * User docs that explain how and why to use Tesseract's automatic differentiation algorithms and operations, with emphasis on scientific machine learning and differentiable physics: [Differentiable physics programming in Tesseract Core docs](https://docs.pasteurlabs.ai/projects/tesseract-core/latest/content/introduction/differentiable-programming.html)

I'm sorry, but when I look at that page, I don't see anything about scientific machine learning or differentiable physics. Machine learning is mentioned at the very top as a motivation, but there's nothing that makes this clear. And "physics" doesn't appear in that page at all.  I think this is a case where the documentation makes a lot of assumptions about the reader and their knowledge.

> * and of course more getting added to the main repo and community forum every week.
>
> We will address your other comments and push the updates ASAP this week.
