+++ docs/content/examples/meshstats.md
@@ -1,11 +1,0 @@
-# Mes Stats
-
-## Context
-TODO -- Missing readme
-
-## Example Tesseract (examples/meshstats)
-
-
-```{literalinclude} ../../../examples/meshstats/tesseract_api.py
-:language: python
-```
