### 2025-05-15 a4ba40c
Directive comments deprecated
