### 2025-05-20 a162ea2
chore: bump codecov/codecov-action from 5.4.2 to 5.4.3 in the actions group (#169)
Bumps the actions group with 1 update:
[codecov/codecov-action](https://github.com/codecov/codecov-action).

Updates `codecov/codecov-action` from 5.4.2 to 5.4.3
<details>
<summary>Release notes</summary>
<p><em>Sourced from <a
href="https://github.com/codecov/codecov-action/releases">codecov/codecov-action's
releases</a>.</em></p>
<blockquote>
<h2>v5.4.3</h2>
<h2>What's Changed</h2>
<ul>
<li>fix: OIDC on forks by <a
href="https://github.com/joseph-sentry"><code>@​joseph-sentry</code></a>
in <a
href="https://red
[…]
