### 2025-05-21 20e4207
feat: check subclass of input- and output schema dynamically in runtime.cli.check to allow for transitive subclass checking
