### 2025-05-21 4f616b7
fix: resolve paths of local relative dependencies
