### 2025-05-23 e8a8d5b
fix: apidoc CLI call used container ID in place of container object to retrieve host port (#172)
#### Relevant issue or PR

#### Description of changes
apidoc call failed due to using a container ID in place of a Container
object when retrieving the doc's host port:

```
│ /usr/local/lib/python3.12/dist-packages/tesseract_core/sdk/cli.py:563 in apidoc                  │
│                                                                                                  │
│   560 │   try:                                                                                   │
│   561 │   │   project_i
[…]
