### 2025-05-23 4050449
Fix arm64 test
