+++ tests/endtoend_tests/test_examples.py
@@ -139,15 +139,9 @@ TEST_CASES = {
                 payload={
                     "inputs": {
                         "vertices": encode_array(
-                            [
-                                [0.0, 666.0, 0.0],
-                                [1.0, 0.0, 0.0],
-                                [0.0, 1.0, 0.0],
-                                [1.0, 1.0, 0.0],
-                                [0.5, 0.5, 1.0],
-                            ]
+                            [[1.0, 2.0, 3.0], [4.0, 5.0, 6.0], [7.0, 8.0, 9.0]]
                         )
-                    },
+                    }
                 },
                 output_contains_pattern="pv_mesh",
             ),
