+++ examples/localdependency/README.md
@@ -1,3 +1,0 @@
-# Build a Tesseract with Local Dependencies
-
-See: [Read the docs](https://docs.pasteurlabs.ai/projects/tesseract-core/howto/)
