### 2025-05-23 f7e9a2a
Fix arm64 tesseract
