+++ examples/pyvista-arm64/tesseract_requirements.txt
@@ -1,3 +1,3 @@
 # Add Python requirements like this:
 # numpy==1.18.1
-pyvista==0.44.2
+# pyvista==0.44.2
