+++ docs/build/html/_downloads/0b714a8159fdb9aa6b7a7f207f411f90/create-vm-azure.sh
@@ -1,0 +1,176 @@
+#!/usr/bin/env bash
+# vim: filetype=bash
+
+set -euo pipefail
+
+# Populate these values with your infrastructure
+REPOSITORY="cr-tesseract-vms.azurecr.io"
+RESOURCE_GROUP="rg-tesseract-vms"
+VNET="vn-tesseract-vms"
+SUBNET="sn-tesseract-vms"
+
+
+error () {
+	echo "$*"
+	exit 1
+}
+
+USAGE="Usage: $0 [OPTION] TESSERACT
+
+TESSERACT must be the Tesseract Image name without the repository, e.g.
+\"meshstats:v1.0.0\".
+
+--vm-size SIZE  defaults to Standard_F2s_v2 (2vCPUs, 4GiB RAM). Other options
+                are described in
+                https://azure.microsoft.com/en-gb/pricing/details/virtual-machines/linux/#pricing
+                Standard_NC4as_T4_v3 is a size with 1 Nvidia T4 GPU.
+
+--gpu           Add this flag to install Nvidia drivers and CUDA toolkit.
+
+--ip-address IP Skip creating the VM and instead run the script in the given IP
+                address.
+"
+
+[[ $# -gt 0 ]] || error "${USAGE}"
+
+# Default values
+baseos_image="Ubuntu2204"
+vm_size="Standard_F2s_v2"
+
+
+# Parse command line arguments/flags
+POSITIONAL_ARGS=()
+while [[ $# -gt 0 ]]; do
+	case $1 in
+		--gpu)
+			GPU=YES
+			shift # past argument
+			;;
+		-i|--ip-address)
+			ip="${2}"
+			shift # past argument
+			shift # past value
+			;;
+		-s|--vm-size)
+			vm_size="${2}"
+			shift # past argument
+			shift # past value
+			;;
+		-h|--help)
+			error "${USAGE}"
+			;;
+		*) # positional arg
+			POSITIONAL_ARGS+=("${1}")
+			shift
+			;;
+	esac
+done
+set -- "${POSITIONAL_ARGS[@]}" # restore positional parameters
+
+if [[ -z ${1+x} ]]; then
+	 echo -e "Missing Tesseract Image\n"
+	 error "${USAGE}"
+fi
+
+tesseract="${1}"
+TESSERACT_NAME=$(echo "${tesseract}" | cut -d ':' -f 1)
+ad_user=$(az ad signed-in-user show --query userPrincipalName \
+          | tr -d '"' \
+          | cut -f 1 -d '@')
+vm_name="${ad_user}-tesseract-${TESSERACT_NAME}-${RANDOM}"
+
+echo -e "Tagging Tesseract image"
+tesseract_image="${REPOSITORY}/${ad_user}/${tesseract}"
+docker image tag "${tesseract}" "${tesseract_image}"
+
+echo -e "Pushing Tesseract image ${tesseract_image}"
+docker push "${tesseract_image}"
+
+
+if [[ -z ${ip+x} ]]; then
+	echo -e "\n\nCreating VM ${vm_name} of instance type ${vm_size}"
+	vm_info=$(az vm create --resource-group "${RESOURCE_GROUP}" \
+			       --subnet "${SUBNET}" \
+			       --vnet-name "${VNET}" \
+			       --generate-ssh-keys \
+			       --public-ip-address '' \
+			       --enable-secure-boot false \
+			       --output json \
+			       --enable-auto-update true \
+			       --admin-username "tessie" \
+			       --name "${vm_name}" \
+			       --image "${baseos_image}" \
+			       --size "${vm_size}")
+
+	ip=$(echo "${vm_info}" | jq -r .privateIpAddress)
+else
+	# shellcheck disable=SC2250
[…]
