+++ docs/build/html/_sources/content/examples/example_gallery.md.txt
@@ -1,0 +1,60 @@
+Example Gallery
+====================
+
+```{toctree}
+:name: example-gallery
+:caption: Example Gallery
+:maxdepth: 2
+:hidden:
+
+arm64
+localdependency
+packagedata
+helloworld
+dataloader
+meshstats
+vectoradd
+```
+
+
+::::{grid} 2
+   :gutter: 2
+
+   :::{grid-item-card} Hello World
+      :link: helloworld.html
+
+      A simple hello world Tesseract.
+   :::
+   :::{grid-item-card} Package Data
+      :link: packagedata.html
+
+      A guide on including local data into Tesseract package.
+   :::
+   :::{grid-item-card} Pyvista on Arm64
+      :link: arm64.html
+
+      A guide on how to set up `tesseract_config` for pyvista on arm64.
+   :::
+   :::{grid-item-card} Local Dependencies
+      :link: localdependency.html
+
+      A guide on installing local dependencies (`cowsay`) into a Tesseract.
+   :::
+   :::{grid-item-card} Mesh Stats
+      :link: meshstats.html
+
+      Tesseract using mesh stats.
+   :::
+   :::{grid-item-card} Data Loader
+      :link: dataloader.html
+
+     Tesseract using dataloader.
+   :::
+   :::{grid-item-card} Vector Add
+      :link: vectoradd.html
+
+     Tesseract doing vector add.
+   :::
+
+
+::::
