+++ docs/build/html/_sources/content/examples/meshstats.md.txt
@@ -1,0 +1,11 @@
+# Mes Stats
+
+## Context
+TODO -- Missing readme
+
+## Example Tesseract (examples/meshstats)
+
+
+```{literalinclude} ../../../examples/meshstats/tesseract_api.py
+:language: python
+```

[…]
