### 2025-05-26 bcf0d31
feat(cli): add serve --no-compose and other missing cli options (#161)
#### Relevant issue or PR
Fixes #156 

#### Description of changes
- Add `tesseract serve --no-compose` to serve Tesseracts via `docker run
--detach` without `docker compose`.
- Changed the notion of "project ID": what was previously strictly a
Docker Compose ID is now either a compose ID or a container name
(`teardown` / `ps` operate on both).
- Add `tesseract serve --num-workers` and `--propagate-tracebacks` while
at it (missing from the CLI).
- Some smaller bug fixes in port handling, where Tesseracts could
sometimes ran
[…]
