### 2025-05-26 ef1e26c
Update tesseract_core/sdk/cli.py
Co-authored-by: Alessandro Angioi <alessandro.angioi@simulation.science>
