### 2025-05-26 98caf85
feat(sdk): make docker executable and build args configurable (#162)
#### Relevant issue or PR
Fixes #154 

#### Description of changes
- Add `tesseract_core.sdk.config` module (copy-paste from
`tesseract_core.runtime.config`).
- Prefix all config envvars with `TESSERACT_` automatically (i.e., we
now use `TESSERACT_DEBUG=1` instead of `DEBUG=1`).
- Introduce config vars for Docker executable and build args.
- Use new build args config to cache Docker builds on GHA. (I don't know
yet if this actually speeds up tests)

#### Testing done
CI, manual

#### License

- [x] By submitting this pull req
[…]
