+++ tesseract_core/sdk/docker_client.py
@@ -6,18 +6,29 @@
 import json
 import logging
 import os
-import re
 import shlex
 import subprocess
 from dataclasses import dataclass
 from pathlib import Path
 from textwrap import indent
+from typing import Literal
+
+from tesseract_core.sdk.config import get_config
 
 logger = logging.getLogger("tesseract")
 
 
 # store a reference to list, which is shadowed by some function names below
 list_ = list
+
+
+def _get_executable(program: Literal["docker", "docker-compose"]) -> tuple[str, ...]:
+    config = get_config()
+    if program == "docker":
+        return config.docker_executable
+    if program == "docker-compose":
+        return config.docker_compose_executable
+    raise ValueError(f"Unknown program: {program}")
 
 
 @dataclass
@@ -32,9 +43,13 @@ class Image:
     @classmethod
     def from_dict(cls, json_dict: dict) -> "Image":
         """Create an Image object from a json dictionary."""
+        image_id = json_dict.get("Id", None)
+        if image_id and not image_id.startswith("sha256:"):
+            # Some container engines (e.g., Podman) do not prefix the ID with sha256:
+            image_id = f"sha256:{image_id}"
         return cls(
-            id=json_dict.get("Id", None),
-            short_id=json_dict.get("Id", [])[:19],
+            id=image_id,
+            short_id=image_id[:19] if image_id else None,
             tags=json_dict.get("RepoTags", None),
             attrs=json_dict,
         )
@@ -95,26 +110,21 @@ class Images:
         if not image_id_or_name:
             raise ValueError("Image name cannot be empty.")
 
-        def is_image_id(s: str) -> bool:
-            """Check if string is image name or id by checking if it's sha256 format."""
-            return bool(re.fullmatch(r"(sha256:)?[a-fA-F0-9]{12,64}", s))
+        def _sanitize_image_id(image_id: str) -> str:
+            """Sanitize image id by removing sha256 prefix."""
+            if image_id.startswith("sha256:"):
+                return image_id[len("sha256:") :]
+            return image_id
 
-        if ":" not in image_id_or_name:
-            # Check if image param is a name or id so we can append latest tag if needed
-            if not is_image_id(image_id_or_name):
-                image_id_or_name = image_id_or_name + ":latest"
-            else:
-                # If image_id_or_name is an image id, we need to get the full id
-                # by prepending sha256
-                image_id_or_name = "sha256:" + image_id_or_name
+        image_id_or_name = _sanitize_image_id(image_id_or_name)
         images = Images.list(tesseract_only=tesseract_only)
 
         # Check for both name and id to find the image
         # Tags may be prefixed by repository url
         for image_obj in images:
             if (
-                image_obj.id == image_id_or_name
-                or image_obj.short_id == image_id_or_name
+                _sanitize_image_id(image_obj.id) == image_id_or_name
+                or _sanitize_image_id(image_obj.short_id) == image_id_or_name
                 or Images._tag_exists(image_id_or_name, image_obj.tags)
             ):
                 return image_obj
@@ -140,9 +150,10 @@ class Images:
         Params:
             image: The image name or id to remove.
         """
+        docker = _get_executable("docker")
         try:
             res = subprocess.run(
-                ["docker", "rmi", image, "--force"],
+                [*docker, "rmi", image, "--force"],
                 check=True,
                 capture_output=True,
                 text=True,
@@ -165,8 +176,10 @@ class Images:
         Returns:
             The buildx command as a list of strings.
         """
+        docker = _get_executable("docker")
+        extra_args = get_config().docker_build_args
         build_cmd = [
-            "docker",
+            *docker,
             "buildx",
             "build",
             "--load",
@@ -174,6 +187,8 @@ class Images:
             tag,
             "--file",
             str(dockerfile),
+            *extra_args,
+            "--",
[…]
