### 2025-05-26 8673eba
feat(sdk): make docker executable and build args configurable (#162)
Fixes #154

- Add `tesseract_core.sdk.config` module (copy-paste from
`tesseract_core.runtime.config`).
- Prefix all config envvars with `TESSERACT_` automatically (i.e., we
now use `TESSERACT_DEBUG=1` instead of `DEBUG=1`).
- Introduce config vars for Docker executable and build args.
- Use new build args config to cache Docker builds on GHA. (I don't know
yet if this actually speeds up tests)

CI, manual

- [x] By submitting this pull request, I confirm that my contribution is
made under the terms of the [Apache 2.0
license
[…]
