### 2025-05-26 7d05cb1
Try to use sh and see what breaks
