+++ .github/workflows/run_tests.yml
@@ -191,6 +191,13 @@ jobs:
 
     runs-on: ${{ matrix.arch == 'x64' && matrix.os == 'ubuntu-24.04' && 'ubuntu-24.04' || matrix.arch == 'arm' && matrix.os == 'ubuntu-24.04' && 'linux-arm64-ubuntu2404'}}
 
+    env:
+      TESSERACT_DOCKER_BUILD_ARGS: "--cache-from type=gha --cache-to type=gha,mode=max"
+
+    defaults:
+      run:
+        shell: bash -el {0}
+
     steps:
       - name: Set up Git repository
         uses: actions/checkout@v4
