### 2025-05-26 025aa22
ci: add CLA Assistant
Signed-off-by: Heitor Pascoal de Bittencourt <heitorpbittencourt@gmail.com>
