### 2025-05-26 726114c
chore: remove mention of DCO from PR template
Signed-off-by: Heitor Pascoal de Bittencourt <heitorpbittencourt@gmail.com>
