### 2025-05-26 7caf35c
chore: mention CLA in contributing guide
