+++ .github/workflows/cla.yml
@@ -1,4 +1,5 @@
 name: "CLA Assistant"
+
 on:
   issue_comment:
     types: [created]
@@ -16,12 +17,10 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - name: "CLA Assistant"
-        if: (github.event.comment.body == 'recheck' || github.event.comment.body == 'I have read the CLA Document and I hereby sign the CLA') || github.event_name == 'pull_request_target'
+        if: (github.event.comment.body == 'recheck' || github.event.comment.body == 'I have read the CLA Document and I hereby sign the CLA') || github.event_name == 'pull_request_target')
         uses: contributor-assistant/github-action@v2.6.1
         env:
-          # the below token should have repo scope and must be manually added by you in the repository's secret
-          # This token is required only if you have configured to store the signatures in a remote repository/organization
-          PERSONAL_ACCESS_TOKEN: ${{ secrets.PERSONAL_ACCESS_TOKEN }}
+          PERSONAL_ACCESS_TOKEN: ${{ secrets.CLA_PAT }}
         with:
           path-to-signatures: 'signatures/tesseract-core/version1/cla.json'
           path-to-document: 'https://github.com/pasteurlabs/pasteur-oss-cla/blob/main/README.md'
@@ -31,5 +30,5 @@ jobs:
 
           remote-organization-name: 'pasteurlabs'
           remote-repository-name: 'pasteur-oss-cla-signatures'
-          create-file-commit-message: 'Creating file for storing CLA Signatures'
+          create-file-commit-message: 'Creating file for storing CLA Signatures for tesseract-core'
           signed-commit-message: '$contributorName has signed the CLA in $owner/$repo#$pullRequestNo'
