### 2025-05-26 ac78797
add dependabot to allowlist
