### 2025-05-27 08b6041
doc: Remove how-to guides froms sidebar (#177)
#### Description of changes
Remove howto guides link from doc sidebar

#### Testing done
make html

#### License

- [x ] By submitting this pull request, I confirm that my contribution
is made under the terms of the [Apache 2.0
license](https://pasteurlabs.github.io/tesseract/LICENSE).
- [ x] I sign the Developer Certificate of Origin below by adding my
name and email address to the `Signed-off-by` line.

<details>
<summary><b>Developer Certificate of Origin</b></summary>

```text
Developer Certificate of Origin
Version 1.1

Copyright (C) 2004, 20
[…]
