### 2025-05-27 59b4786
Move dynamic checks to runtime.core and add existence checks for required schemas and endpoints.
