+++ tests/sdk_tests/test_api_parse.py
@@ -175,16 +175,14 @@ def test_optional_signature_check(
 
 
 def test_schema_parent_class_is_checked(
-    tmp_path, valid_tesseract_api, valid_tesseract_config, caplog
+    tmp_path, valid_tesseract_api, valid_tesseract_config
 ):
-    tesseract_api = valid_tesseract_api.replace("(BaseModel)", "")
-    _write_tesseract_api_to_file(tesseract_api, tmp_path)
-    _write_tesseract_config_to_file(valid_tesseract_config, tmp_path)
+    for schema in ("InputSchema", "OutputSchema"):
+        tesseract_api = valid_tesseract_api.replace(f"{schema}(BaseModel)", schema)
+        _write_tesseract_api_to_file(tesseract_api, tmp_path)
+        _write_tesseract_config_to_file(valid_tesseract_config, tmp_path)
 
-    with caplog.at_level(logging.WARNING):
-        validate_tesseract_api(tmp_path)
-        for schema in ("InputSchema", "OutputSchema"):
-            assert (
-                f"{schema} does not directly derive from pydantic.BaseModel (got: None)"
-                in caplog.text
-            )
+        with pytest.raises(
+            ValidationError, match=f"{schema} must define a parent class"
+        ):
+            validate_tesseract_api(tmp_path)
