### 2025-05-27 60f6eb0
remove abstract_eval from required endpoints
