### 2025-05-27 8d609dd
ci(cla): don't use tesseract-core dir for signatures
