+++ .github/workflows/cla.yml
@@ -23,7 +23,7 @@ jobs:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
           PERSONAL_ACCESS_TOKEN: ${{ secrets.CLA_PAT }}
         with:
-          path-to-signatures: 'signatures/tesseract-core/version1/cla.json'
+          path-to-signatures: 'signatures/version1/cla.json'
           path-to-document: 'https://github.com/pasteurlabs/pasteur-oss-cla/blob/main/README.md'
           # Branch should not be protected
           branch: 'main'
