### 2025-05-27 2486d29
also remove abstract_eval from test
