### 2025-05-27 cc9a73d
fixed remaining issues (print statement, unused import, error message)
