### 2025-05-27 047c474
add link to tesseract showcase
