### 2025-05-27 a027a49
feat: more comprehensive validation of input and output schema during `tesseract-runtime check` (#170)
#### Description of changes
Move the input- and output schema subclass check to `runtime.cli.check`
to allow for dynamic transitive subclass checking.
- Only produce a warning when parsing AST in
`sdk.api_parse.validate_tesseract_api`
- Check if input- and output schema are subclasses of
"pydantic.BaseModel" in `sdk.api_parse.validate_tesseract_api`
- Add corresponding tests

#### Testing done
- Tested changes locally
- Introduced new tests to cover new subclass checking

#### License

- [x] 
[…]
