### 2025-05-28 477108d
Remove unnecessary license ack
Signed-off-by: Heitor Pascoal de Bittencourt <heitorpbittencourt@gmail.com>
