### 2025-05-28 f47a59e
try using docker inspect instead of comparing strings
