+++ tesseract_core/sdk/cli.py
@@ -34,7 +34,6 @@ from .docker_client import (
     BuildError,
     CLIDockerClient,
     Container,
-    ContainerError,
     Image,
     ImageNotFound,
 )
