### 2025-05-29 7a3843b
Add data assimilation tesseract
