+++ docs/content/examples/example_gallery.md
@@ -4,50 +4,63 @@ Example Gallery
 ```{toctree}
 :name: example-gallery
 :caption: Example Gallery
-:maxdepth: 2
 :hidden:
 
-arm64
-localdependency
-packagedata
-helloworld
-dataloader
-vectoradd
+superb-tutorials/lorenz.ipynb
 ```
 
+
+## Building Blocks Tesseracts
+
+This is a gallery of Tesseracts that end at the `build` stage of the Tesseract lifecycle, and can act as building blocks for your own Tesseract. You can find these Tesseracts in the `examples` directory of the codebase.
 
 ::::{grid} 2
    :gutter: 2
 
    :::{grid-item-card} Hello World
-      :link: helloworld.html
+      :link: building-blocks/helloworld.html
 
       A simple hello world Tesseract.
    :::
    :::{grid-item-card} Package Data
-      :link: packagedata.html
+      :link: building-blocks/packagedata.html
 
       A guide on including local data into Tesseract package.
    :::
    :::{grid-item-card} Vector Add
-      :link: vectoradd.html
+      :link: building-blocks/vectoradd.html
 
      Tesseract doing vector add.
    :::
    :::{grid-item-card} Pyvista on Arm64
-      :link: arm64.html
+      :link: building-blocks/arm64.html
 
       A guide on how to set up `tesseract_config` for pyvista on arm64.
    :::
    :::{grid-item-card} Local Dependencies
-      :link: localdependency.html
+      :link: building-blocks/localdependency.html
 
       A guide on installing local dependencies (`cowsay`) into a Tesseract.
    :::
    :::{grid-item-card} Data Loader
-      :link: dataloader.html
+      :link: building-blocks/dataloader.html
 
      Tesseract that loads in data samples from a folder.
+   :::
+
+::::
+
+## Beyond the Build
+
+These are some more advanced Tesseracts that showcase the true capabilities of a Tesseract, beyond just the `build` stage of the Tesseract lifecycle.
+
+::::{grid} 2
+   :gutter: 2
+
+   :::{grid-item-card} 4D-Variational Data Assimilation for Chaotic Dynamical System
+      :link: superb-tutorials/lorenz.html
+
+      Near Real-Time Sequential Data Assimilation for State Prediction in a Low-Dimensional Surrogate of the Two-Scale Lorenz-96 System.
    :::
 
 ::::
