+++ advanced_examples/lorenz/requirements.txt
@@ -1,0 +1,2 @@
+optax
+matplotlib
