+++ requirements.txt
@@ -41,6 +41,15 @@ rich==14.0.0
 shellingham==1.5.4
     # via typer
 typer==0.15.2
+myst-nb==0.19.4
+jupyter-cache==0.6.4
+jupyter-client==8.7.0
+jupyter-core==5.7.2
+jupyter-server==2.13.1
+jupyterlab-pygments==0.2.2
+nbformat==5.9.3
+nbconvert==7.12.0
+notebook==7.1.3
     # via tesseract-core
 typing-extensions==4.13.2
     # via

[…]
