### 2025-05-29 8827df4
Try building julia-mlp on ci
