+++ .github/workflows/cla.yml
@@ -3,7 +3,7 @@ name: "CLA Assistant"
 on:
   issue_comment:
     types: [created]
-  pull_request_target:
+  pull_request:
     types: [opened,closed,synchronize]
 
 jobs:
