+++ .github/workflows/cla.yml
@@ -11,11 +11,11 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - name: "CLA Assistant"
-        if: startsWith(github.event.comment.body, '@PasteurBot') || github.event_name == 'pull_request'
+        if: startsWith(github.event.comment.body, '@PasteurBot') || github.event.comment.body == 'recheck' || github.event_name == 'pull_request'
         uses: contributor-assistant/github-action@v2.6.1
         env:
           GITHUB_TOKEN: ${{ secrets.PL_PASTEURBOT_PAT_PUBLIC }}
-          PERSONAL_ACCESS_TOKEN: ${{ secrets.CLA_PAT }}
+          # PERSONAL_ACCESS_TOKEN: ${{ secrets.CLA_PAT }}
         with:
           path-to-signatures: 'signatures/version1/cla.json'
           path-to-document: 'https://github.com/pasteurlabs/pasteur-oss-cla/blob/main/README.md'
@@ -27,9 +27,9 @@ jobs:
           create-file-commit-message: 'Creating file for storing CLA Signatures for tesseract-core'
           signed-commit-message: '$contributorName has signed the CLA in $owner/$repo#$pullRequestNo'
           custom-notsigned-prcomment: |
-            Thank you for your contribution!
+            Thank you for your PR!
             Like many open-source projects, we ask that you sign our Contributor License Agreement before we can accept your contribution.
-            You can do so by commenting `@PasteurBot I have read the CLA Document and I hereby sign the CLA` on this PR.
+            You can do so by commenting the following on this pull request:
           custom-allsigned-prcomment: |
             All contributors have signed the CLA! Your contribution is now eligible for merging.
           custom-pr-sign-comment: |
