+++ .github/workflows/cla.yml
@@ -17,20 +17,23 @@ jobs:
           GITHUB_TOKEN: ${{ secrets.PL_PASTEURBOT_PAT_PUBLIC }}
           PERSONAL_ACCESS_TOKEN: ${{ secrets.PL_PASTEURBOT_PAT_PUBLIC }}
         with:
-          path-to-signatures: 'signatures/version1/cla.json'
-          path-to-document: 'https://github.com/pasteurlabs/pasteur-oss-cla/blob/main/README.md'
-          # Branch should not be protected
-          branch: 'main'
-          allowlist: 'dependabot[bot],PasteurBot'
+          allowlist: 'dependabot[bot],PasteurBot,github-actions[bot]'
+
           remote-organization-name: 'pasteurlabs'
           remote-repository-name: 'pasteur-oss-cla-signatures'
+          path-to-signatures: 'signatures/version1/cla.json'
+          branch: 'main'
+
           create-file-commit-message: 'Creating file for storing CLA Signatures for tesseract-core'
           signed-commit-message: '$contributorName has signed the CLA in $owner/$repo#$pullRequestNo'
-          custom-notsigned-prcomment: |
+          custom-notsigned-prcomment: >
             Thank you for your PR!
-            Like many open-source projects, we ask that you sign our Contributor License Agreement before we can accept your contribution.
+            Like many open-source projects, we ask that you sign our
+            [Contributor License Agreement](https://github.com/pasteurlabs/pasteur-oss-cla/blob/main/README.md)
+            before we can accept your contribution.
+
             You can do so by commenting the following on this pull request:
           custom-allsigned-prcomment: |
-            All contributors have signed the CLA! Your contribution is now eligible for merging.
+            ✅ All contributors have signed the CLA! Your contribution is now eligible for merging.
           custom-pr-sign-comment: |
             @PasteurBot I have read the CLA Document and I hereby sign the CLA
