+++ .github/workflows/cla.yml
@@ -3,7 +3,7 @@ name: "CLA Assistant"
 on:
   issue_comment:
     types: [created]
-  pull_request:
+  pull_request_target:
     types: [opened,closed,synchronize]
 
 jobs:
@@ -11,7 +11,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - name: "CLA Assistant"
-        if: startsWith(github.event.comment.body, '@PasteurBot') || github.event.comment.body == 'recheck' || github.event_name == 'pull_request'
+        if: startsWith(github.event.comment.body, '@PasteurBot') || github.event.comment.body == 'recheck' || github.event_name == 'pull_request_target'
         uses: contributor-assistant/github-action@v2.6.1
         env:
           GITHUB_TOKEN: ${{ secrets.PL_PASTEURBOT_PAT_PUBLIC }}
@@ -27,7 +27,7 @@ jobs:
           create-file-commit-message: 'Creating file for storing CLA Signatures for tesseract-core'
           signed-commit-message: '$contributorName has signed the CLA in $owner/$repo#$pullRequestNo'
           custom-notsigned-prcomment: >
-            ### ⚠️ CLA signature required ⚠️
+            #### ⚠️ CLA signature required ⚠️
 
             Thank you for your PR, we really appreciate it!
             Like many open-source projects, we ask that you sign our
@@ -35,7 +35,7 @@ jobs:
             before we can accept your contribution.
             You can do so by commenting the following on this pull request:
           custom-allsigned-prcomment: >
-            ### 🎉 CLA signatures confirmed 🎉
+            #### 🎉 CLA signatures confirmed 🎉
 
             All contributors have signed the [Contributor License Agreement](https://github.com/pasteurlabs/pasteur-oss-cla/blob/main/README.md).
           custom-pr-sign-comment: |
