### 2025-05-29 02daf1c
fix julia tesseract
