### 2025-05-29 3742a96
overhaul error handling; do not detach containers when running
