+++ docs/content/examples/example_gallery.md
@@ -4,52 +4,47 @@ Example Gallery
 ```{toctree}
 :name: example-gallery
 :caption: Example Gallery
-:maxdepth: 2
 :hidden:
 
-arm64
-localdependency
-packagedata
-helloworld
-dataloader
-vectoradd
+superb-tutorials/tesseracting_surrogates/tesseract_linear_ellipse
+superb-tutorials/optimization_with_surrogates/linear_ellipse_optimization
 ```
 
 
 ## Building Blocks Tesseracts
 
-This is a gallery of some extremely basic features of Tesseracts that can act as building blocks for your own Tesseract.
+This is a gallery of some extremely basic features of Tesseracts that can act as building blocks for your own Tesseract. You can find these Tesseracts in the `examples` directory of the codebase.
 
 ::::{grid} 2
    :gutter: 2
 
    :::{grid-item-card} Hello World
-      :link: helloworld.html
+      :link: building-blocks/helloworld.html
 
       A simple hello world Tesseract.
    :::
    :::{grid-item-card} Package Data
-      :link: packagedata.html
+      :link: building-blocks/packagedata.html
 
       A guide on including local data into Tesseract package.
    :::
    :::{grid-item-card} Vector Add
-      :link: vectoradd.html
+      :link: building-blocks/vectoradd.html
 
      Tesseract doing vector add.
    :::
    :::{grid-item-card} Pyvista on Arm64
-      :link: arm64.html
+      :link: building-blocks/arm64.html
 
       A guide on how to set up `tesseract_config` for pyvista on arm64.
    :::
    :::{grid-item-card} Local Dependencies
-      :link: localdependency.html
+      :link: building-blocks/localdependency.html
 
       A guide on installing local dependencies (`cowsay`) into a Tesseract.
    :::
    :::{grid-item-card} Data Loader
-      :link: dataloader.html
+      :link: building-blocks/dataloader.html
 
      Tesseract that loads in data samples from a folder.
    :::
@@ -64,14 +59,30 @@ Some more advanced Tesseracts that showcase the true capabilities of a Tesseract
    :gutter: 2
 
    :::{grid-item-card} Interoperable Autodiff Engines, aka Tesseracted Surrogates
-      :link: https://legendary-adventure-z4k3rek.pages.github.io/content/tutorials/tesseracting_surrogates/tesseract_linear_ellipse.html
+      :link: superb-tutorials/tesseracting_surrogates/tesseract_linear_ellipse.html
 
       Packaging any trained differentiable ROM or ML surrogate in a standardized, universal container with standardized deployment.
    :::
-   :::{grid-item-card} Package Design Optimization with Tesseracts
-      :link: https://legendary-adventure-z4k3rek.pages.github.io/content/tutorials/optimization_with_surrogates/linear_ellipse_optimization.html
+   :::{grid-item-card} Engineering Design Optimization with Tesseracts
+      :link: superb-tutorials/optimization_with_surrogates/linear_ellipse_optimization.html
 
       Using any differentiable ROM or ML surrogate within an optimization framework without worrying about the actual emulator-optimizer interactions
    :::
+   :::{grid-item-card} _Coming soon_: CAxML Data Generation Pipelines
+      :link: superb-tutorials/caxml_datagen_automations/caxml_data_automations.html
+
+      Inverse design loops, continual learning surrogates, and more Tesseract applications with CAE simulation software tool in the loop
+   :::
+   :::{grid-item-card} _Coming soon_: Bridging Simulation and Reality with Differentiable Programming
+      :link: superb-tutorials/sim_to_real/sim_to_real.html
+
+      Generating hybrid models that bridge simulations with real experimental data
[…]
