+++ docs/content/examples/superb-tutorials/micro_to_macro/micro_to_macro.md
@@ -1,0 +1,12 @@
+---
+orphan: true
+---
+
+# Integrating Micro and Macro Scale Simulations
+
+```{figure} micro_to_macro.png
+:alt: micro_to_macro
+:width: 900px
+```
+
+Coming soon: demo applications for accelerating multiscale simulations!
