+++ docs/content/examples/superb-tutorials/tesseracting_surrogates/tesseracting_surrogates/setting_up_tesseract_requirements_julia.md
@@ -1,0 +1,21 @@
+# Setting up the `tesseract_requirements.txt` file for Julia surrogate
+
+The default generated `tesseract_requirements.txt` looks as follows.
+
+```
+# Tesseract requirements file
+# Generated by tesseract 0.3.0+6.g9ece6f8 on 2024-08-22T09:00:59.002494
+
+# Add Python requirements like this:
+# numpy==1.18.1
+
+# This may contain private dependencies via SSH URLs:
+# git+ssh://git@github.com/username/repo.git@branch
+# (use `tesseract build --forward_ssh_agent` to grant the builder access to your SSH keys)
+```
+For this demo, we specifiy the dependencies of the <span class="product">Tesseract</span>.
+
+```
+numpy==2.0.0
+juliacall==0.9.20
+```
