+++ docs/content/examples/superb-tutorials/tesseracting_surrogates/tesseracting_surrogates/tesseracting_surrogates.md
@@ -1,0 +1,14 @@
+# Tesseracting Surrogates
+
+```{toctree}
+:maxdepth: 1
+tesseract_quickstart.md
+setting_up_tesseract_api_julia.md
+setting_up_config_yaml_julia.md
+setting_up_tesseract_requirements_julia.md
+building_julia_tesseract.md
+rest_api_julia_tesseract.md
+setting_up_tesseract_jax.md
+building_jax_tesseract.md
+rest_api_jax_tesseract.md
+```
