+++ docs/content/examples/superb-tutorials/optimization_with_surrogates/linear_ellipse_optimization.md
@@ -1,0 +1,16 @@
+# Design Optimization using Tesseracted Surrogates
+
+This tutorial demonstrates how to use <span class="product">Tesseracts</span> within a customized optimization solution. It shows how to interact with <span class="product">Tesseracts</span> to obtain surrogate outputs and gradients and it demonstrates
+one possible way to formulate optimization problems with the Pasteur Labs <span class="product">Plasmosis</span> engine. This tutorial concludes with an example of Multi-Objective Optimization use two separate <span class="product">Tesseracts</span>.
+
+```{figure} optimization_with_surrogates.png
+:alt: linear-elasticity-ellipse
+:width: 1200px
+```
+
+## Contents
+```{toctree}
+:maxdepth: 1
+
+linear_ellipse_optimization_page.md
+```

[…]
