### 2025-05-30 72ea23e
remove emojis
