### 2025-06-02 4c25752
fix: Overhaul docker client for better podman compatibility and better error handling (#178)
#### Relevant issue or PR
Fixes #155 

#### Description of changes
- Be less opinionated about how image IDs are represented and just use
whatever the used Docker implementation returns. This for example means
we now both accept `sha256:<hash>` or `<hash>` as image ID, and just
pass it to the Docker CLI unmodified.
- Use `docker inspect <image_id>` during `images.get` and
`containers.get` instead of looping over images / containers and
comparing IDs.
- Do not detach containers during `tesseract run` (p
[…]
