+++ tesseract_core/sdk/docker_client.py
@@ -35,8 +35,8 @@ def _get_executable(program: Literal["docker", "docker-compose"]) -> tuple[str, 
 class Image:
     """Image class to wrap Docker image details."""
 
-    id: str
-    short_id: str
+    id: str | None
+    short_id: str | None
     tags: list[str] | None
     attrs: dict
 
@@ -44,12 +44,17 @@ class Image:
     def from_dict(cls, json_dict: dict) -> "Image":
         """Create an Image object from a json dictionary."""
         image_id = json_dict.get("Id", None)
-        if image_id and not image_id.startswith("sha256:"):
-            # Some container engines (e.g., Podman) do not prefix the ID with sha256:
-            image_id = f"sha256:{image_id}"
+        short_id = None
+        if image_id:
+            if image_id.startswith("sha256:"):
+                short_id = image_id[:19]
+            else:
+                # Some container engines (e.g., Podman) do not prefix IDs with sha256
+                short_id = image_id[:12]
+
         return cls(
             id=image_id,
-            short_id=image_id[:19] if image_id else None,
+            short_id=short_id,
             tags=json_dict.get("RepoTags", None),
             attrs=json_dict,
         )
@@ -57,34 +62,6 @@ class Image:
 
 class Images:
     """Namespace for functions to interface with Tesseract docker images."""
-
-    @staticmethod
-    def _tag_exists(image_name: str, tags: list_) -> bool:
-        """Helper function to check if image name exists in the list of tags.
-
-        Specially handling has to be done to achieve unfuzzy substring matching, i.e.
-        if image tag is foo/bar/image, we need to return true for foo/bar/image, bar/image,
-        and image, but not ar/image.
-
-        There is no equivalent in docker-py.
-
-        Params:
-            image_name: The image name to check.
-            tags: The list of tags to check against.
-
-        Returns:
-            True if the image name exists in the list of tags, False otherwise.
-        """
-        if ":" not in image_name:
-            image_name += ":latest"
-
-        image_name_parts = image_name.strip("/").split("/")
-        for tag in tags:
-            tag_parts = tag.strip("/").split("/")
-            # Check if the last parts of the tag matches the subportion that image name specifies
-            if tag_parts[-len(image_name_parts) :] == image_name_parts:
-                return True
-        return False
 
     @staticmethod
     def get(image_id_or_name: str | bytes, tesseract_only: bool = True) -> Image:
@@ -110,26 +87,26 @@ class Images:
         if not image_id_or_name:
             raise ValueError("Image name cannot be empty.")
 
-        def _sanitize_image_id(image_id: str) -> str:
-            """Sanitize image id by removing sha256 prefix."""
-            if image_id.startswith("sha256:"):
-                return image_id[len("sha256:") :]
-            return image_id
+        docker = _get_executable("docker")
+        try:
+            result = subprocess.run(
+                [*docker, "inspect", image_id_or_name, "--type", "image"],
+                check=True,
+                capture_output=True,
+                text=True,
+            )
+            json_dict = json.loads(result.stdout)
+        except subprocess.CalledProcessError as ex:
+            raise ImageNotFound(f"Image {image_id_or_name} not found.") from ex
+        if not json_dict:
+            raise ImageNotFound(f"Image {image_id_or_name} not found.")
 
-        image_id_or_name = _sanitize_image_id(image_id_or_name)
-        images = Images.list(tesseract_only=tesseract_only)
-
-        # Check for both name and id to find the image
-        # Tags may be prefixed by repository url
-        for image_obj in images:
-            if (
-                _sanitize_image_id(image_obj.id) == image_id_or_name
-                or _sanitize_image_id(image_obj.short_id) == image_id_or_name
-                or Images._tag_exists(image_id_or_name, image_obj.tags)
-            ):
-                return image_obj
-
[…]
