+++ tesseract_core/sdk/engine.py
@@ -681,7 +681,7 @@ def _create_docker_compose_template(
         else:
             gpu_settings = f"device_ids: {gpus}"
 
-    for image_id, port in zip(image_ids, ports, strict=True):
+    for i, (image_id, port) in enumerate(zip(image_ids, ports, strict=True)):
         service = {
             "name": f"{image_id.split(':')[0]}-{_id_generator()}",
             "image": image_id,
@@ -691,6 +691,7 @@ def _create_docker_compose_template(
             "environment": {
                 "TESSERACT_DEBUG": "1" if debug else "0",
             },
+            "debugpy_port": 5678 + i,
         }
 
         services.append(service)
