### 2025-06-02 c171277
Add debugpy dep
