+++ tesseract_core/runtime/serve.py
@@ -89,6 +89,12 @@ def create_rest_api(api_module: ModuleType) -> FastAPI:
 
 def serve(host: str, port: int, num_workers: int) -> None:
     """Start the REST API."""
+    config = get_config()
+    if config.debug:
+        import debugpy
+
+        debugpy.listen(("0.0.0.0", 5678))
+
     uvicorn.run(
         "tesseract_core.runtime.app_http:app", host=host, port=port, workers=num_workers
     )
