+++ CONTRIBUTING.md
@@ -160,9 +160,17 @@ or by adding a `BREAKING CHANGE:` trailer to the message.
 The Tesseract project follows [semantic versioning](https://semver.org).
 
 
-## Changelog
+## Release process
 
-This project [changelog](CHANGELOG.md) is generated by
-[git-cliff](https://git-cliff.org). The changelog is generated by running
-`git cliff --output CHANGELOG.md`. To bump the semantic version of the project
-when generating the changelog: `git cliff --output CHANGELOG.md --bump`.
+Releases are done via GitHub Actions, which automatically build the release
+artifacts and publish them to the [GitHub Releases](https://github.com/pasteurlabs/tesseract-core/releases) page. To create a new release, follow these steps:
+
+1. Make sure the code is in a good state, all tests pass, and the documentation is up to date.
+2. Trigger a new release action through the GitHub UI: []() # TODO!
+   This opens a new pull request with the release notes and the version number.
+3. Add any additional release notes to the pull request.
+4. In the meantime, you can add more commits to `main` which will automatically
+   be included in the release.
+5. Once the pull request is ready, merge it into `main`.
+
+That's it! GitHub Actions will automatically release the new version.
