+++ .github/workflows/release.yml
@@ -1,0 +1,145 @@
+name: Handle release process
+
+on:
+  # Case 1: manual trigger -> initiate release process
+  workflow_dispatch:
+
+  # Case 2: release PR is being updated -> update changelog
+  pull_request:
+    types: [opened, reopened, synchronize]  # Trigger on PR events
+    branches:
+      - main
+
+  # Case 3: release PR is merged -> create a new release
+  push:
+    branches:
+      - main
+
+
+jobs:
+  trigger-pr:
+    runs-on: ubuntu-latest
+    if: github.event_name == 'workflow_dispatch'
+
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v2
+
+      - name: Set up Python
+        uses: actions/setup-python@v4
+        with:
+          python-version-file: "pyproject.toml"
+
+      - name: Install git-cliff
+        run: |
+          pip install git-cliff
+
+      - name: Generate changelog
+        id: generate_changelog
+        run: |
+          git-cliff --output CHANGELOG.md --bump
+          new_version=$(git-cliff --bumped-version)
+          echo "new_version=$new_version" >> $GITHUB_OUTPUT
+
+      - name: Trigger release PR
+        uses: peter-evans/create-pull-request@v3
+        with:
+          title: "Release Changelog"
+          body: >
+            This PR contains the generated changelog for the release
+            ${{ steps.generate_changelog.outputs.new_version }}.
+
+            ⚠️ **Merging this PR will immediately trigger a new release**. ⚠️
+
+            To specify additional release notes, please edit this comment after the
+            following line.
+
+            ---
+
+            <!-- release-notes -->
+          commit-message: "chore: update changelog"
+          branch: "bot/release/${{ steps.generate_changelog.outputs.new_version }}"
+          base: "main"
+          token: ${{ secrets.PL_PASTEURBOT_PAT_PUBLIC }}
+
+  update-pr:
+    runs-on: ubuntu-latest
+    if: github.event_name == 'pull_request' && startsWith(github.head_ref, 'refs/heads/bot/release/')
+
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v2
+        with:
+          token: ${{ secrets.PL_PASTEURBOT_PAT_PUBLIC }}
+
+      - name: Set up Python
+        uses: actions/setup-python@v4
+        with:
+          python-version-file: "pyproject.toml"
+
+      - name: Install git-cliff
+        run: |
+          pip install git-cliff
+
+      - name: Generate changelog
+        id: generate_changelog
+        run: |
+          git-cliff --output CHANGELOG.md --bump
+          new_version=$(git-cliff --bumped-version)
+          echo "new_version=$new_version" >> $GITHUB_OUTPUT
+
+      - name: Update PR with changelog
+        run: |
+          git add CHANGELOG.md
+          git config --global user.name "PasteurBot"
+          git config --global user.email "hello@simulation.science"
+          git commit -m "chore: update changelog before release"
+          git push origin HEAD:${{ github.head_ref }}
+
+  release:
+    runs-on: ubuntu-latest
+    if: github.event_name == 'push' && startsWith(github.ref, 'refs/heads/main/')
+
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v2
+        with:
+          token: ${{ secrets.PL_PASTEURBOT_PAT_PUBLIC }}
+
+      - name: Set up Python
+        uses: actions/setup-python@v4
+        with:
+          python-version-file: "pyproject.toml"
+
+      - name: Install git-cliff
+        run: |
+          pip install git-cliff
+
+      - name: Get new version
+        id: get_version
+        run: |
[…]
