### 2025-06-03 19987a9
add ability to configure host IP
