### 2025-06-03 6aead27
add host ip attribute to containers; replace localhost with 127.0.0.1
