### 2025-06-03 8d38edc
add test and fix compose handling
