### 2025-06-04 b1df8e6
ci: Release automation (#181)
Add an action to initiate and handle release process (generation of
changelogs and release notes).
