+++ CHANGELOG.md
@@ -2,137 +2,4 @@
 
 All notable changes to this project will be documented in this file.
 
-## [0.9.0] - 2025-05-02
-
-### Features
-
-- [**breaking**] Remove docker_py usage in favor of custom client that uses Docker CLI (#33)
-- *(sdk)* Allow users to serve Tesseracts using multiple worker processes (#135)
-
-### Documentation
-
-- Update quickstart (#144)
-
-## [0.8.5] - 2025-04-24
-
-### Bug Fixes
-
-- Fixed typos in jax recipe (#134)
-- *(sdk)* Various improvements to SDK UX (#136)
-
-## [0.8.4] - 2025-04-17
-
-### Features
-
-- Allow creating tesseract objects from python modules (#122)
-- Also allow passing an imported module to Tesseract.from_tesseract_api (#130)
-
-## [0.8.3] - 2025-04-14
-
-### Bug Fixes
-
-- Fix Tesseract SDK decoding and error handling (#123)
-
-## [0.8.2] - 2025-04-11
-
-### Features
-
-- Add requirement provider config to build Tesseracts from conda env specs (#54)
-- Introduce debug mode for served Tesseracts to propagate tracebacks to clients (#111)
-
-### Bug Fixes
-
-- Ensure Ubuntu-based base images work as expected; change default to vanilla Debian (#115)
-- Enable users to opt-in to allowing extra fields in Tesseract schemas by setting `extra="allow"` (#117)
-- Meshstats `abstract_eval` (#120)
-
-## [0.8.1] - 2025-03-28
-
-### Bug Fixes
-
-- Pydantic 2.11.0 compatibility (hotfix) (#106)
-
-## [0.8.0] - 2025-03-27
-
-### Features
-
-- Implement check_gradients runtime command (#72)
-- [**breaking**] Validate endpoint argument names before building (#95)
-
-### Bug Fixes
-
-- OpenAPI schema failure for differentiable arrays with unknown shape (#100)
-- Prevent silent conversion of float array to int (#96)
-- Use fixed uid/gid 5000:5000 for all tesseracts (#102)
-- Use uid 1000 instead of 5000 (#104)
-
-### Refactor
-
-- Unpack endpoint payload (#80)
-
-### Documentation
-
-- Dependencies and user privileges (#91)
-
-## [0.7.4] - 2025-03-20
-
-### Features
-
-- Friendlier error messages when input validation fails (#71)
-- Pytorch initialize template (#53)
-- Add `diffable` field to input/output json schemas (#82)
-- Add stdout output for tesseract build (#87)
-
-### Documentation
-
-- Various docs nits + UX fixes (#85)
-
-## [0.7.3] - 2025-03-13
-
-### Features
-
-- Raise proper error (`ValidationError`) for invalid inputs (#67)
-- Add `abstract_eval` method to `tesseract_core.Tesseract` (#76)
-
-### Bug Fixes
-
-- Jax template now uses equinox `filter_jit` to allow non-array inputs (#56)
-- Added pip as dependency (#58)
-- Issue #74 (#75)
-
-### Documentation
-
-- Updated comments in jax recipe and docs on Differentiable flag (#65)
-
-## [0.7.2] - 2025-02-27
-
-### Bug Fixes
-
-- Validate ShapeDType in abstract-eval schemas (#40)
-- Resolve paths before passing volumes to docker (#48)
-- Dangling Tesseracts in e2e tests (#51)
-- Sanitize error output (#52)
-
-### Documentation
-
-- Python API for Julia example (#37)
-- Fix links again (#49)
-
-## [0.7.1] - 2025-02-26
-
-### Bug Fixes
-
[…]
