+++ .github/workflows/bump_lockfile.yml
@@ -13,6 +13,8 @@ jobs:
     steps:
       - name: Set up Git repository
         uses: actions/checkout@v4
+        with:
+            token: ${{ secrets.PL_PASTEURBOT_PAT_PUBLIC }}
 
       - name: Install uv
         uses: astral-sh/setup-uv@v6
@@ -46,7 +48,7 @@ jobs:
         if: steps.git-diff.outputs.update_done == 'true'
         uses: peter-evans/create-pull-request@v7
         with:
-            token: ${{ secrets.GITHUB_TOKEN }}
+            token: ${{ secrets.PL_PASTEURBOT_PAT_PUBLIC }}
             commit-message: Update dependencies
             title: "chore: 📦 Update dependencies"
             branch: _bot/update-deps
@@ -55,5 +57,3 @@ jobs:
             body: |
               This PR updates the lockfile to the latest versions of the dependencies.
               Please review the changes and merge when ready.
-
-              To trigger CI checks, please close and reopen this PR.
