+++ .github/workflows/release.yml
@@ -19,7 +19,11 @@ jobs:
 
     steps:
       - name: Checkout code
-        uses: actions/checkout@v2
+        uses: actions/checkout@v4
+        with:
+          token: ${{ secrets.PL_PASTEURBOT_PAT_PUBLIC }}
+          fetch-depth: 0
+          fetch-tags: true
 
       - name: Set up Python
         uses: actions/setup-python@v4
@@ -40,19 +44,15 @@ jobs:
       - name: Trigger release PR
         uses: peter-evans/create-pull-request@v3
         with:
-          title: "Release Changelog"
-          body: >
-            This PR contains the generated changelog for the release
-            ${{ steps.generate_changelog.outputs.new_version }}.
+          title: "chore: release v${{ steps.generate_changelog.outputs.new_version }}"
+          body: |
+            This PR contains the generated changelog for the release ${{ steps.generate_changelog.outputs.new_version }}.
 
             ⚠️ **Merging this PR will immediately trigger a new release**. ⚠️
 
-            To specify additional release notes, please edit this comment after the
-            following line.
+            To specify additional release notes, please edit this comment after the following line.
 
             ---
-
-            <!-- release-notes -->
           commit-message: "chore: update changelog"
           branch: "bot/release/${{ steps.generate_changelog.outputs.new_version }}"
           base: "main"
@@ -64,9 +64,11 @@ jobs:
 
     steps:
       - name: Checkout code
-        uses: actions/checkout@v2
+        uses: actions/checkout@v4
         with:
           token: ${{ secrets.PL_PASTEURBOT_PAT_PUBLIC }}
+          fetch-depth: 0
+          fetch-tags: true
 
       - name: Set up Python
         uses: actions/setup-python@v4
@@ -98,9 +100,11 @@ jobs:
 
     steps:
       - name: Checkout code
-        uses: actions/checkout@v2
+        uses: actions/checkout@v4
         with:
           token: ${{ secrets.PL_PASTEURBOT_PAT_PUBLIC }}
+          fetch-depth: 0
+          fetch-tags: true
 
       - name: Set up Python
         uses: actions/setup-python@v4
@@ -123,7 +127,7 @@ jobs:
           changelog=$(git cliff --bump --latest --strip all)
 
           custom_notes="${{ github.event.head_commit.message }}"
-          custom_notes=$(echo "$custom_notes" | awk '/<!-- release-notes -->/{flag=1;next} flag')
+          custom_notes=$(echo "$custom_notes" | awk '/^---$/,/^---$/ {next} {print}')
 
           touch /tmp/notes.md
           echo "# Release Notes for v${{ steps.get_version.outputs.new_version }}" > /tmp/notes.md
