+++ .github/workflows/release.yml
@@ -44,7 +44,7 @@ jobs:
       - name: Trigger release PR
         uses: peter-evans/create-pull-request@v3
         with:
-          title: "chore: release v${{ steps.generate_changelog.outputs.new_version }}"
+          title: "chore: release ${{ steps.generate_changelog.outputs.new_version }}"
           body: |
             This PR contains the generated changelog for the release ${{ steps.generate_changelog.outputs.new_version }}.
 
@@ -88,11 +88,17 @@ jobs:
 
       - name: Update PR with changelog
         run: |
-          git add CHANGELOG.md
           git config --global user.name "PasteurBot"
           git config --global user.email "hello@simulation.science"
-          git commit -m "chore: update changelog before release"
-          git push origin HEAD:${{ github.head_ref }}
+
+          git add CHANGELOG.md
+
+          if git diff --cached --quiet; then
+            echo "No changes to commit"
+          else
+            git commit -m "chore: update changelog before release"
+            git push origin HEAD:${{ github.head_ref }}
+          fi
 
   release:
     runs-on: ubuntu-latest
