### 2025-06-04 994afb2
feat: Add ability to configure host IP during `tesseract serve` (#185)
#### Relevant issue or PR
n/a

#### Description of changes
- Serve on `127.0.0.1` (localhost) by default when using
`tesseract-runtime serve`, to avoid accidentally exposing Tesseracts to
the local or public internet when using `tesseract-runtime` directly.
- Add capability to configure host IP when serving containers via
`tesseract serve`.

To-do:
- [x] Also expose in Python API
- [x] Add tests

#### Testing done
CI + manual
