### 2025-06-04 efabf2b
try using gh cli; head_ref -> ref
