### 2025-06-04 f00bfa3
chore: Fix minor release CI issues (#192)
#### Relevant issue or PR
Release CI job still having a few issues, this hopefully fixes it.

#### Description of changes
* Remove spurious v in pr title
* only push if there is something in the changelog which changed
(otherwise job fails, see
https://github.com/pasteurlabs/tesseract-core/actions/runs/15439467957/job/43453632741#step:6:19
)

---------

Co-authored-by: Dion Häfner <dion.haefner@simulation.science>
