+++ .github/workflows/release.yml
@@ -42,21 +42,29 @@ jobs:
           echo "new_version=$new_version" >> $GITHUB_OUTPUT
 
       - name: Trigger release PR
-        uses: peter-evans/create-pull-request@v3
-        with:
-          title: "chore: release v${{ steps.generate_changelog.outputs.new_version }}"
-          body: |
-            This PR contains the generated changelog for the release ${{ steps.generate_changelog.outputs.new_version }}.
+        run: |
+          git config --global user.name "PasteurBot"
+          git config --global user.email "hello@simulation.science"
 
-            ⚠️ **Merging this PR will immediately trigger a new release**. ⚠️
+          git checkout -b "bot/release/${{ steps.generate_changelog.outputs.new_version }}"
+          git add CHANGELOG.md
+          git commit -m "chore: update changelog"
 
-            To specify additional release notes, please edit this comment after the following line.
+          pr_body=$(cat <<EOF
+          This PR contains the generated changelog for the release ${{ steps.generate_changelog.outputs.new_version }}.
 
-            ---
-          commit-message: "chore: update changelog"
-          branch: "bot/release/${{ steps.generate_changelog.outputs.new_version }}"
-          base: "main"
-          token: ${{ secrets.PL_PASTEURBOT_PAT_PUBLIC }}
+          ⚠️ **Merging this PR will immediately trigger a new release**. ⚠️
+
+          To specify additional release notes, please edit this comment after the following line.
+
+          ---
+          EOF
+          )
+
+          gh pr create \
+            --title "chore: release ${{ steps.generate_changelog.outputs.new_version }}" \
+            --body "$pr_body" \
+            --base "main"
 
   update-pr:
     runs-on: ubuntu-latest
@@ -88,11 +96,17 @@ jobs:
 
       - name: Update PR with changelog
         run: |
-          git add CHANGELOG.md
           git config --global user.name "PasteurBot"
           git config --global user.email "hello@simulation.science"
-          git commit -m "chore: update changelog before release"
-          git push origin HEAD:${{ github.head_ref }}
+
+          git add CHANGELOG.md
+
+          if git diff --cached --quiet; then
+            echo "No changes to commit"
+          else
+            git commit -m "chore: update changelog before release"
+            git push origin HEAD:${{ github.ref }}
+          fi
 
   release:
     runs-on: ubuntu-latest
@@ -142,5 +156,4 @@ jobs:
             --notes-file "${{ steps.release_notes.outputs.release_note_file }}" \
             --draft \
             --target ${{ github.sha }} \
-            --repo ${{ github.repository }} \
             --latest
