### 2025-06-04 dea6600
chore: update changelog before release
