+++ .github/workflows/release.yml
@@ -11,11 +11,16 @@ on:
   # Case 1: manual trigger -> initiate release process
   workflow_dispatch:
 
+  # Case 2: release PR is being updated -> update changelog
   push:
     branches:
-      # Case 2: release PR is being updated -> update changelog
       - 'bot/release/*'
-      # Case 3: release PR is merged -> create a new release
+
+  # Case 3: release PR is merged -> create a new release
+  pull_request:
+    types:
+      - closed
+    branches:
       - main
 
 env:
@@ -121,13 +126,14 @@ jobs:
 
   release:
     runs-on: ubuntu-latest
-    if: github.event_name == 'push' && startsWith(github.ref, 'refs/heads/main/')
+    if: github.event_name == 'pull_request' && github.event.pull_request.merged == true && startsWith(github.event.pull_request.head.ref, 'bot/release/')
 
     steps:
       - name: Checkout code
         uses: actions/checkout@v4
         with:
           token: ${{ secrets.PL_PASTEURBOT_PAT_PUBLIC }}
+          ref: ${{ github.event.pull_request.merge_commit_sha }}
           fetch-depth: 0
           fetch-tags: true
 
@@ -151,7 +157,7 @@ jobs:
         run: |
           changelog=$(git cliff --bump --latest --strip all)
 
-          custom_notes="${{ github.event.head_commit.message }}"
+          custom_notes="${{ github.event.pull_request.body }}"
           custom_notes=$(echo "$custom_notes" | awk '/^---$/,/^---$/ {next} {print}')
 
           touch /tmp/notes.md
@@ -166,5 +172,5 @@ jobs:
             --title "v${{ steps.get_version.outputs.new_version }}" \
             --notes-file "${{ steps.release_notes.outputs.release_note_file }}" \
             --draft \
-            --target ${{ github.sha }} \
+            --target ${{ github.event.pull_request.merge_commit_sha }} \
             --latest
