### 2025-06-04 7799228
chore: cosmetics on release notes
