+++ .github/workflows/release.yml
@@ -155,21 +155,30 @@ jobs:
       - name: Assemble release notes
         id: release_notes
         run: |
-          changelog=$(git cliff --bump --latest --strip all)
+          changelog=$(git cliff --latest --strip all)
+          # strip off first line (which is the version number)
+          changelog=$(echo "$changelog" | sed '1d')
 
           custom_notes="${{ github.event.pull_request.body }}"
-          custom_notes=$(echo "$custom_notes" | awk '/^---$/,/^---$/ {next} {print}')
+          # Keep only the part after the first '---' line
+          custom_notes=$(echo "$custom_notes" | awk '/^---$/ {found=1; next} found && !/^---$/ {print}')
+          # Strip leading / trailing whitespace
+          custom_notes=$(echo "$custom_notes" | sed 's/^[ \t]*//;s/[ \t]*$//')
 
+          # Create the release notes file
           touch /tmp/notes.md
-          echo "# Release Notes for v${{ steps.get_version.outputs.new_version }}" > /tmp/notes.md
-          printf "%s\n\n" "$custom_notes" >> /tmp/notes.md
-          printf "## Changelog\n\n%s\n" "$changelog" >> /tmp/notes.md
+          echo "# Release ${{ steps.get_version.outputs.new_version }}" > /tmp/notes.md
+          if [ ! -z "$custom_notes" ]; then
+            printf "%s\n\n" "$custom_notes" >> /tmp/notes.md
+          fi
+          printf "## Changes\n%s\n" "$changelog" >> /tmp/notes.md
+
           echo "release_note_file=/tmp/notes.md" >> $GITHUB_OUTPUT
 
       - name: Create new release
         run: |
-          gh release create "v${{ steps.get_version.outputs.new_version }}" \
-            --title "v${{ steps.get_version.outputs.new_version }}" \
+          gh release create "${{ steps.get_version.outputs.new_version }}" \
+            --title "${{ steps.get_version.outputs.new_version }}" \
             --notes-file "${{ steps.release_notes.outputs.release_note_file }}" \
             --draft \
             --target ${{ github.event.pull_request.merge_commit_sha }} \
