### 2025-06-04 253f11b
chore: more cosmetics on release notes
