+++ .github/workflows/release.yml
@@ -133,9 +133,16 @@ jobs:
         uses: actions/checkout@v4
         with:
           token: ${{ secrets.PL_PASTEURBOT_PAT_PUBLIC }}
-          ref: ${{ github.event.pull_request.merge_commit_sha }}
+          ref: ${{ github.event.pull_request.base.ref }}
           fetch-depth: 0
           fetch-tags: true
+
+      - name: Ensure we're on the correct commit
+        run: |
+          if [ "$(git rev-parse HEAD)" != "${{ github.event.pull_request.merge_commit_sha }}" ]; then
+            echo "Error: The current commit does not match the merge commit SHA."
+            exit 1
+          fi
 
       - name: Set up Python
         uses: actions/setup-python@v4
@@ -157,13 +164,13 @@ jobs:
         run: |
           changelog=$(git cliff --latest --strip all)
           # strip off first line (which is the version number)
-          changelog=$(echo "$changelog" | sed '1d')
+          changelog=$(printf "%s" "$changelog" | sed '1d')
 
           custom_notes="${{ github.event.pull_request.body }}"
           # Keep only the part after the first '---' line
-          custom_notes=$(echo "$custom_notes" | awk '/^---$/ {found=1; next} found && !/^---$/ {print}')
+          custom_notes=$(printf "%s" "$custom_notes" | awk '/^---$/ {found=1; next} found && !/^---$/ {print}')
           # Strip leading / trailing whitespace
-          custom_notes=$(echo "$custom_notes" | sed 's/^[ \t]*//;s/[ \t]*$//')
+          custom_notes=$(printf "%s" "$custom_notes" | sed 's/^[ \t]*//;s/[ \t]*$//')
 
           # Create the release notes file
           touch /tmp/notes.md
@@ -172,6 +179,8 @@ jobs:
             printf "%s\n\n" "$custom_notes" >> /tmp/notes.md
           fi
           printf "## Changes\n%s\n" "$changelog" >> /tmp/notes.md
+
+          cat /tmp/notes.md
 
           echo "release_note_file=/tmp/notes.md" >> $GITHUB_OUTPUT
 
