### 2025-06-04 7fe559c
chore: more cosmetics on release notes
