### 2025-06-04 bc3995d
disable other tests for now
