+++ requirements.txt
@@ -2,9 +2,9 @@
 #    uv export --frozen --color never --no-emit-project --no-hashes
 annotated-types==0.7.0
     # via pydantic
-certifi==2025.4.26
+certifi==2025.1.31
     # via requests
-charset-normalizer==3.4.2
+charset-normalizer==3.4.1
     # via requests
 click==8.1.8
     # via typer
@@ -22,7 +22,7 @@ mdurl==0.1.2
     # via markdown-it-py
 numpy==2.2.4
     # via tesseract-core
-pip==25.1.1
+pip==25.0.1
     # via tesseract-core
 pydantic==2.11.0
     # via tesseract-core
@@ -41,24 +41,15 @@ rich==14.0.0
 shellingham==1.5.4
     # via typer
 typer==0.15.2
-myst-nb==0.19.4
-jupyter-cache==0.6.4
-jupyter-client==8.7.0
-jupyter-core==5.7.2
-jupyter-server==2.13.1
-jupyterlab-pygments==0.2.2
-nbformat==5.9.3
-nbconvert==7.12.0
-notebook==7.1.3
     # via tesseract-core
-typing-extensions==4.13.2
+typing-extensions==4.13.0
     # via
     #   pydantic
     #   pydantic-core
     #   rich
     #   typer
     #   typing-inspection
-typing-inspection==0.4.1
+typing-inspection==0.4.0
     # via pydantic
-urllib3==2.4.0
+urllib3==2.3.0
     # via requests
