+++ .github/workflows/release.yml
@@ -160,15 +160,14 @@ jobs:
           changelog=$(printf "%s" "$changelog" | sed '1d')
 
           # Get custom notes from the PR body
-          read -r -d '' custom_notes <<EOF
-          ${{ github.event.pull_request.body }}
-          EOF
+          custom_notes="${{ github.event.pull_request.body }}""
+          echo "$custom_notes"
           # Keep only the part after the first '---' line
           custom_notes=$(echo "$custom_notes" | sed -n '/^---$/,$p' | sed '1d')
-          echo $custom_notes
+          echo "$custom_notes"
           # Strip leading / trailing whitespace
           custom_notes=$(printf "%s" "$custom_notes" | sed 's/^[ \t]*//;s/[ \t]*$//')
-          echo $custom_notes
+          echo "$custom_notes"
 
           # Create the release notes file
           touch /tmp/notes.md
