+++ .github/workflows/release.yml
@@ -160,7 +160,7 @@ jobs:
           changelog=$(printf "%s" "$changelog" | sed '1d')
 
           # Get custom notes from the PR body
-          custom_notes="${{ github.event.pull_request.body }}""
+          custom_notes="${{ github.event.pull_request.body }}"
           echo "$custom_notes"
           # Keep only the part after the first '---' line
           custom_notes=$(echo "$custom_notes" | sed -n '/^---$/,$p' | sed '1d')
