### 2025-06-04 800ba1e
Update git workflow description
