+++ .github/workflows/run_advanced_examples.yml
@@ -21,8 +21,7 @@ jobs:
         python-version: ["3.10"]
 
         example:
-          - simple
-          - cfd
+          - lorenz
 
       fail-fast: false
 
