+++ requirements.txt
@@ -2,9 +2,9 @@
 #    uv export --frozen --color never --no-emit-project --no-hashes
 annotated-types==0.7.0
     # via pydantic
-certifi==2025.4.26
+certifi==2025.1.31
     # via requests
-charset-normalizer==3.4.2
+charset-normalizer==3.4.1
     # via requests
 click==8.1.8
     # via typer
@@ -22,7 +22,7 @@ mdurl==0.1.2
     # via markdown-it-py
 numpy==2.2.4
     # via tesseract-core
-pip==25.1.1
+pip==25.0.1
     # via tesseract-core
 pydantic==2.11.0
     # via tesseract-core
@@ -42,14 +42,14 @@ shellingham==1.5.4
     # via typer
 typer==0.15.2
     # via tesseract-core
-typing-extensions==4.13.2
+typing-extensions==4.13.0
     # via
     #   pydantic
     #   pydantic-core
     #   rich
     #   typer
     #   typing-inspection
-typing-inspection==0.4.1
+typing-inspection==0.4.0
     # via pydantic
-urllib3==2.4.0
+urllib3==2.3.0
     # via requests
