### 2025-06-05 88bfbb9
chore: fix release action for real (#199)
Co-authored-by: PasteurBot <hello@simulation.science>
