+++ .github/workflows/release.yml
@@ -154,16 +154,19 @@ jobs:
 
       - name: Assemble release notes
         id: release_notes
+        env:
+          PR_BODY: ${{ github.event.pull_request.body }}
         run: |
           changelog=$(git cliff --unreleased --bump --latest --strip all)
           # strip off first line (which is the version number)
           changelog=$(printf "%s" "$changelog" | sed '1d')
 
           # Get custom notes from the PR body
-          custom_notes="${{ github.event.pull_request.body }}"
-          echo $custom_notes
+          custom_notes="$PR_BODY"
+          # Fix line endings
+          custom_notes=$(printf "%s" "$custom_notes" | tr -d '\r')
           # Keep only the part after the first '---' line
-          custom_notes=$(printf "%s" "$custom_notes" | awk '/^---$/ {found=1; next} found && !/^---$/ {print}')
+          custom_notes=$(echo "$custom_notes" | sed -n '/^---$/,$p' | sed '1d')
           # Strip leading / trailing whitespace
           custom_notes=$(printf "%s" "$custom_notes" | sed 's/^[ \t]*//;s/[ \t]*$//')
 
@@ -175,8 +178,6 @@ jobs:
           fi
           printf "## Changes\n%s\n" "$changelog" >> /tmp/notes.md
 
-          cat /tmp/notes.md
-
           echo "release_note_file=/tmp/notes.md" >> $GITHUB_OUTPUT
 
       - name: Create new release
@@ -184,6 +185,5 @@ jobs:
           gh release create "${{ steps.get_version.outputs.new_version }}" \
             --title "${{ steps.get_version.outputs.new_version }}" \
             --notes-file "${{ steps.release_notes.outputs.release_note_file }}" \
-            --draft \
             --target ${{ github.event.pull_request.merge_commit_sha }} \
             --latest
