+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-## [0.9.1] - 2025-06-04
+## [0.9.1] - 2025-06-05
 
 ### Features
 
