+++ .github/workflows/release.yml
@@ -164,9 +164,10 @@ jobs:
           changelog=$(printf "%s" "$changelog" | sed '1d')
 
           # Get custom notes from the PR body
-          custom_notes="${{ env.PR_BODY }}"
+          # custom_notes="${{ env.PR_BODY }}"
           # Fix line endings
-          custom_notes=$(printf "%s" "$custom_notes" | tr -d '\r')
+          custom_notes=$(printf "%s" "$PR_BODY" | tr -d '\r')
+          printf "%s" "$custom_notes"
           # Keep only the part after the first '---' line
           custom_notes=$(echo "$custom_notes" | sed -n '/^---$/,$p' | sed '1d')
           # Strip leading / trailing whitespace
