+++ .github/workflows/release.yml
@@ -164,9 +164,9 @@ jobs:
           changelog=$(printf "%s" "$changelog" | sed '1d')
 
           # Get custom notes from the PR body
+          custom_notes="$PR_BODY"
           # Fix line endings
-          printf "%s" "$PR_BODY"
-          custom_notes=$(printf "%s" "$PR_BODY" | tr -d '\r')
+          custom_notes=$(printf "%s" "$custom_notes" | tr -d '\r')
           printf "%s" "$custom_notes"
           # Keep only the part after the first '---' line
           custom_notes=$(echo "$custom_notes" | sed -n '/^---$/,$p' | sed '1d')
