+++ .github/workflows/release.yml
@@ -18,8 +18,8 @@ on:
 
   # Case 3: release PR is merged -> create a new release
   pull_request:
-    # types:
-    #   - closed
+    types:
+      - closed
     branches:
       - main
 
@@ -126,7 +126,7 @@ jobs:
 
   release:
     runs-on: ubuntu-latest
-    # if: github.event_name == 'pull_request' && github.event.pull_request.merged == true && startsWith(github.event.pull_request.head.ref, 'bot/release/')
+    if: github.event_name == 'pull_request' && github.event.pull_request.merged == true && startsWith(github.event.pull_request.head.ref, 'bot/release/')
 
     steps:
       - name: Checkout code
@@ -161,7 +161,7 @@ jobs:
           PR_BODY: ${{ github.event.pull_request.body }}
         run: |
           changelog=$(git-cliff --unreleased --bump --latest --strip all)
-          # strip off first line (which is the version number)
+          # Strip off first line (which is the version number)
           changelog=$(printf "%s" "$changelog" | sed '1d')
 
           # Get custom notes from the PR body
@@ -184,13 +184,12 @@ jobs:
           # Append link to full diff
           echo -e "\n**Full diff**: https://github.com/${{ github.repository_owner }}/${{ github.repository }}/compare/${{ steps.get_version.outputs.current_version }}...${{ steps.get_version.outputs.new_version }}" >> /tmp/notes.md
 
-          cat /tmp/notes.md
           echo "release_note_file=/tmp/notes.md" >> $GITHUB_OUTPUT
 
-      # - name: Create new release
-      #   run: |
-      #     gh release create "${{ steps.get_version.outputs.new_version }}" \
-      #       --title "${{ steps.get_version.outputs.new_version }}" \
-      #       --notes-file "${{ steps.release_notes.outputs.release_note_file }}" \
-      #       --target ${{ github.event.pull_request.merge_commit_sha }} \
-      #       --latest
+      - name: Create new release
+        run: |
+          gh release create "${{ steps.get_version.outputs.new_version }}" \
+            --title "${{ steps.get_version.outputs.new_version }}" \
+            --notes-file "${{ steps.release_notes.outputs.release_note_file }}" \
+            --target ${{ github.event.pull_request.merge_commit_sha }} \
+            --latest
