+++ .github/workflows/run_tests.yml
@@ -1,292 +1,292 @@
-# name: Run test suite
+name: Run test suite
 
-# on:
-#   pull_request:
+on:
+  pull_request:
 
-#   push:
-#     branches:
-#       - main
+  push:
+    branches:
+      - main
 
-# concurrency:
-#   group: ${{ github.workflow }}-${{ github.event.pull_request.number || github.ref }}
-#   cancel-in-progress: true
+concurrency:
+  group: ${{ github.workflow }}-${{ github.event.pull_request.number || github.ref }}
+  cancel-in-progress: true
 
-# jobs:
-#   tests-base:
-#     strategy:
-#       matrix:
-#         os: [ubuntu-latest, windows-latest, macos-latest]
+jobs:
+  tests-base:
+    strategy:
+      matrix:
+        os: [ubuntu-latest, windows-latest, macos-latest]
 
-#         # test with oldest and latest supported Python versions
-#         # NOTE: If bumping the minimum Python version here, also do it in
-#         # ruff.toml, setup.py and other CI files as well.
-#         python-version: ["3.10", "3.12"]
-#         runtime-deps: ["latest"]
+        # test with oldest and latest supported Python versions
+        # NOTE: If bumping the minimum Python version here, also do it in
+        # ruff.toml, setup.py and other CI files as well.
+        python-version: ["3.10", "3.12"]
+        runtime-deps: ["latest"]
 
-#         include:
-#           - os: ubuntu-latest
-#             python-version: "3.12"
-#             runtime-deps: "oldest"
+        include:
+          - os: ubuntu-latest
+            python-version: "3.12"
+            runtime-deps: "oldest"
 
-#       fail-fast: false
+      fail-fast: false
 
-#     runs-on: ${{ matrix.os }}
+    runs-on: ${{ matrix.os }}
 
-#     steps:
-#       - name: Set up Git repository
-#         uses: actions/checkout@v4
+    steps:
+      - name: Set up Git repository
+        uses: actions/checkout@v4
 
-#       - name: Set up WSL (Windows)
-#         uses: Vampire/setup-wsl@v5
-#         if: runner.os == 'Windows'
-#         with:
-#           distribution: Ubuntu-24.04
+      - name: Set up WSL (Windows)
+        uses: Vampire/setup-wsl@v5
+        if: runner.os == 'Windows'
+        with:
+          distribution: Ubuntu-24.04
 
-#       - name: Set up Python
-#         if: runner.os != 'Windows'
-#         uses: actions/setup-python@v5
-#         with:
-#           python-version: ${{ matrix.python-version }}
+      - name: Set up Python
+        if: runner.os != 'Windows'
+        uses: actions/setup-python@v5
+        with:
+          python-version: ${{ matrix.python-version }}
 
-#       - name: Set up Python (Windows)
-#         if: runner.os == 'Windows'
-#         shell: wsl-bash {0}
-#         run: |
-#           python_version=${{ matrix.python-version }}
-#           if [ $python_version == "3.12" ]; then
-#             # Python 3.12 is the default `python3` in Ubuntu 24.04
-#             python_version="3"
-#           fi
+      - name: Set up Python (Windows)
+        if: runner.os == 'Windows'
+        shell: wsl-bash {0}
+        run: |
+          python_version=${{ matrix.python-version }}
+          if [ $python_version == "3.12" ]; then
+            # Python 3.12 is the default `python3` in Ubuntu 24.04
+            python_version="3"
+          fi
 
-#           export DEBIAN_FRONTEND=noninteractive
-#           sudo apt-get update
[…]
