### 2025-06-05 6c9439c
chore: release v0.9.1 (#201)
This PR contains the generated changelog for the release v0.9.1.

⚠️ **Merging this PR will immediately trigger a new release**. ⚠️

To specify additional release notes, please edit this comment after the
following line.

---

This release brings several improvements and bug fixes that make it
easier to work with Tesseracts.

The `tesseract` CLI now supports additional configuration:
- Config for the used Docker executables via
`TESSERACT_DOCKER_EXECUTABLE`, `TESSERACT_DOCKER_COMPOSE_EXECUTABLE`,
and `TESSERACT_DOCKER_BUILD_ARGS` environment variables.
- Config for
[…]
