+++ .github/workflows/release.yml
@@ -182,7 +182,7 @@ jobs:
           printf "## What's Changed\n%s\n" "$changelog" >> /tmp/notes.md
 
           # Append link to full diff
-          echo -e "\n**Full diff**: https://github.com/${{ github.repository_owner }}/${{ github.repository }}/compare/${{ steps.get_version.outputs.current_version }}...${{ steps.get_version.outputs.new_version }}" >> /tmp/notes.md
+          echo -e "\n**Full diff**: https://github.com/${{ github.repository }}/compare/${{ steps.get_version.outputs.current_version }}...${{ steps.get_version.outputs.new_version }}" >> /tmp/notes.md
 
           echo "release_note_file=/tmp/notes.md" >> $GITHUB_OUTPUT
 
