### 2025-06-05 4079c24
simplify release PR creation
