### 2025-06-05 8997bd0
chore: Roll out PasteurBot to all automated actions (#190)
#### Relevant issue or PR
n/a

#### Description of changes
Use @PasteurBot as actor for all automated PRs.

(+ some drive-by changes to fix issues with the release action)

#### Testing done
https://github.com/pasteurlabs/tesseract-core/pull/203

---------

Co-authored-by: Alessandro Angioi <alessandro.angioi@simulation.science>
