### 2025-06-05 a9f3da5
refactor example structure
