+++ docs/content/examples/building-blocks/helloworld.md
@@ -1,15 +1,10 @@
----
-orphan: true
----
-
-# Hello World
+# HelloWorld
 
 ## Context
-This is a basic Hello World Tesseract. See: [#Get Started](#getting-started)
+This is a basic Hello World Tesseract. See [getting-started]
 for more details on this Tesseract.
 
-## Example Tesseract (examples/helloworld)
-
+## Example Tesseract (`examples/helloworld`)
 
 ```{literalinclude} ../../../../examples/helloworld/tesseract_api.py
 :language: python
