+++ docs/content/examples/building-blocks/packagedata.md
@@ -1,8 +1,4 @@
----
-orphan: true
----
-
-# Passing in local data into a Tesseract
+# Packaging local files into a Tesseract
 
 ## Context
 You may need to include local files into your Tesseract, e.g. if we want to load in pretrained model weights, or artifacts that we may need access to. This guide will demonstrate how.
