+++ docs/content/examples/example_gallery.md
@@ -1,66 +1,60 @@
-Example Gallery
-====================
+# Building blocks
 
 ```{toctree}
 :name: example-gallery
 :caption: Example Gallery
+:maxdepth: 2
 :hidden:
-
-superb-tutorials/lorenz.ipynb
+:glob:
+building-blocks/helloworld.md
+building-blocks/vectoradd.md
+building-blocks/packagedata.md
+building-blocks/arm64.md
+building-blocks/localdependency.md
+building-blocks/dataloader.md
 ```
 
+This is a gallery of Tesseract examples that end at the `build` stage of the Tesseract lifecycle, and that can act as starting points to define and build your own Tesseracts.
 
-## Building Blocks Tesseracts
+You can also find these Tesseracts in the `examples` directory of the [code repository](https://github.com/pasteurlabs/tesseract-core).
 
-This is a gallery of Tesseracts that end at the `build` stage of the Tesseract lifecycle, and can act as building blocks for your own Tesseract. You can find these Tesseracts in the `examples` directory of the codebase.
+```{important}
+**Beyond the Build**: The real magic happens long after building a Tesseract. For some example applications that *use* Tesseracts in workflows, check out the [Demo](../demo/demo.md) and [Community Showcase](https://si-tesseract.discourse.group/c/showcase/11).
+```
+
 
 ::::{grid} 2
    :gutter: 2
 
-   :::{grid-item-card} Hello World
+   :::{grid-item-card} HelloWorld
       :link: building-blocks/helloworld.html
 
-      A simple hello world Tesseract.
+      A simple "hello world" Tesseract.
+   :::
+   :::{grid-item-card} VectorAdd
+      :link: building-blocks/vectoradd.html
+
+     Tesseract performing vector addition. Highlighting simple array operations and how to use the Tesseract Python API.
    :::
    :::{grid-item-card} Package Data
       :link: building-blocks/packagedata.html
 
-      A guide on including local data into Tesseract package.
+      A guide on including local files into a built Tesseract.
    :::
-   :::{grid-item-card} Vector Add
-      :link: building-blocks/vectoradd.html
-
-     Tesseract doing vector add.
-   :::
-   :::{grid-item-card} Pyvista on Arm64
+   :::{grid-item-card} Pyvista on ARM64
       :link: building-blocks/arm64.html
 
-      A guide on how to set up `tesseract_config` for pyvista on arm64.
+      A guide showcasing how to use custom build steps to install pyvista within an ARM64 Tesseract.
    :::
    :::{grid-item-card} Local Dependencies
       :link: building-blocks/localdependency.html
 
-      A guide on installing local dependencies (`cowsay`) into a Tesseract.
+      A guide on installing local Python packages into a Tesseract.
    :::
    :::{grid-item-card} Data Loader
       :link: building-blocks/dataloader.html
 
-     Tesseract that loads in data samples from a folder.
-   :::
-
-::::
-
-## Beyond the Build
-
-These are some more advanced Tesseracts that showcase the true capabilities of a Tesseract, beyond just the `build` stage of the Tesseract lifecycle.
-
-::::{grid} 2
-   :gutter: 2
-
-   :::{grid-item-card} 4D-Variational Data Assimilation for Chaotic Dynamical System
-      :link: superb-tutorials/lorenz.html
-
-      Near Real-Time Sequential Data Assimilation for State Prediction in a Low-Dimensional Surrogate of the Two-Scale Lorenz-96 System.
+     Tesseract that loads in data samples from a folder without loading them into memory.
    :::
 
 ::::
