+++ docs/content/demo/demo.md
@@ -1,0 +1,9 @@
+# Demo: 4D-variational data assimilation with Tesseract
+
+```{toctree}
+:caption: Contents
+:maxdepth: 1
+
+data-assimilation-4dvar.ipynb
+lorenz_tesseract.md
+```
