+++ advanced_examples/lorenz/requirements.txt
@@ -1,2 +1,0 @@
-optax
-matplotlib
