### 2025-06-05 981c4e6
rename advanced example -> demo
