+++ docs/content/demo/demo.md
@@ -3,7 +3,24 @@
 ```{toctree}
 :caption: Contents
 :maxdepth: 1
+:hidden:
 
 data-assimilation-4dvar.ipynb
 lorenz_tesseract.md
 ```
+
+(cards-clickable)=
+
+
+:::{card} 4D-Variational Data Assimilation for a Chaotic Dynamical System
+:link: data-assimilation-4dvar.html
+
+Demo using differentiable Tesseracts to implement a 4D-Var data assimilation scheme
+for a chaotic dynamical system
+:::
+
+:::{card} Lorenz Tesseract
+:link: lorenz_tesseract.html
+
+Detailed implementation of the Lorenz Tesseract used in the 4D-Var demo.
+:::
