### 2025-06-05 b38cb69
Allow configuring service names on multi-tesseract network
