+++ tests/sdk_tests/test_engine.py
@@ -207,6 +207,15 @@ def test_serve_tesseracts_invalid_input_args():
         with pytest.raises(ValueError):
             engine.teardown(None)
 
+        with pytest.raises(ValueError):
+            engine.serve(["vectoradd"], ports=["8080", "8081"])
+
+        with pytest.raises(ValueError):
+            engine.serve(["vectoradd"], service_names=["A", "B"])
+
+        with pytest.raises(ValueError):
+            engine.serve(["vectoradd", "vectoradd"], no_compose=True, service_names=["VA1", "VA2"])
+
 
 def test_get_tesseract_images(mocked_docker):
     tesseract_images = engine.get_tesseract_images()
@@ -240,6 +249,10 @@ def test_serve_tesseracts(mocked_docker):
 
     # Serve with gpus
     project_name_multi_tesseract = engine.serve(["vectoradd"], gpus=["1", "3"])
+    assert project_name_multi_tesseract
+
+    # Serve and specify tesseract service names
+    project_name_multi_tesseract = engine.serve(["vectoradd", "vectoradd"], service_names=["VA1", "VA2"])
     assert project_name_multi_tesseract
 
 
