+++ tesseract_core/sdk/cli.py
@@ -468,6 +468,17 @@ def serve(
             ),
         ),
     ] = False,
+    service_names: Annotated[
+        str | None,
+        typer.Option(
+            "--service-names",
+            help=(
+                "Comma-separated list of service names by which each tesseract should be exposed "
+                "in the shared network. "
+                "Tesseracts are reachable from one another at http://{service_name}:8000"
+            ),
+        ),
+    ] = None
 ) -> None:
     """Serve one or more Tesseract images.
 
@@ -491,6 +502,11 @@ def serve(
     else:
         ports = None
 
+    if service_names is not None:
+        service_names_list = service_names.split(',')
+    else:
+        service_names_list = None
+
     try:
         project_id = engine.serve(
             image_names,
@@ -501,6 +517,7 @@ def serve(
             propagate_tracebacks,
             num_workers,
             no_compose,
+            service_names_list,
         )
     except RuntimeError as ex:
         raise UserError(
