### 2025-06-05 4e94819
Allow configuring service names on multi-tesseract network
