### 2025-06-06 c34932b
doc: Add in data assimilation tutorial and refactor example gallery (#200)
#### Description of changes

- Refactored example gallery into building blocks demo
- Added description for example gallery
- Added example for data assimilation
- Changed some stuff in the ipynb for formatting / linter

#### Testing done
Manually regenerated site.. pending ci testing

<img width="779" alt="image"
src="https://github.com/user-attachments/assets/91e422aa-0fa9-4189-a999-08c5003536a2"
/>

---------

Co-authored-by: Dion Häfner <dion.haefner@simulation.science>
Co-authored-by: Alessandro Angioi <alessandro.
[…]
