+++ tesseract_core/runtime/serve.py
@@ -1,10 +1,12 @@
 # Copyright 2025 Pasteur Labs. All Rights Reserved.
 # SPDX-License-Identifier: Apache-2.0
 
+import asyncio
 import inspect
 from functools import wraps
 from types import ModuleType
 from typing import Any, Callable, Union
+from uuid import uuid4
 
 import uvicorn
 from fastapi import FastAPI, Header, Response
@@ -19,6 +21,31 @@ GET_ENDPOINTS = {"input_schema", "output_schema", "health"}
 
 # TODO: make this configurable via environment variable
 DEFAULT_ACCEPT = "application/json"
+
+
+# This function needs to be in global scope, so that it can be offloaded to a separate worker process.
+def _run_api_function_within_worker(
+    func_name: str, picklable_payload_dict: dict
+) -> dict:
+    # Need to load the user-supplied api module and retrieve API endpoints here,
+    # since the worker doesn't inherit local (in the Python sense) definitions from the main process!
+    from tesseract_core.runtime.core import get_tesseract_api
+
+    api_module = get_tesseract_api()
+    endpoints = create_endpoints(api_module)
+
+    endpoint_func = next(
+        (endpoint for endpoint in endpoints if endpoint.__name__ == func_name), None
+    )
+    assert endpoint_func is not None, (
+        f"Function {func_name} not found in endpoints, but worker was tasked with its execution."
+    )
+
+    payload_model = endpoint_func.input_schema.model_validate(
+        picklable_payload_dict
+    )  # TODO: Maybe skip the actual validation here, since it was already done in the main process?
+
+    return endpoint_func(payload=payload_model).model_dump()
 
 
 def create_response(model: BaseModel, accept: str) -> Response:
@@ -44,7 +71,15 @@ def create_rest_api(api_module: ModuleType) -> FastAPI:
     )
     tesseract_endpoints = create_endpoints(api_module)
 
-    def wrap_endpoint(endpoint_func: Callable):
+    from concurrent.futures import ProcessPoolExecutor
+
+    executor = ProcessPoolExecutor(
+        max_workers=4
+    )  # TODO: Adjust the number of workers according to user-configured num_workers!
+
+    open_tasks = {}
+
+    def wrap_endpoint(endpoint_func: Callable, async_endpoint: bool = False):
         endpoints_to_wrap = [
             "apply",
             "jacobian",
@@ -54,13 +89,40 @@ def create_rest_api(api_module: ModuleType) -> FastAPI:
 
         @wraps(endpoint_func)
         async def wrapper(*args: Any, accept: str, **kwargs: Any):
-            result = endpoint_func(*args, **kwargs)
-            return create_response(result, accept)
+            if async_endpoint:
+                new_task_id = str(uuid4())
+                # new_task_id = "1" # TODO: Remove this hardcoded task ID, since it is only for testing purposes!
+
+                assert "payload" in kwargs, (
+                    f"Defining async. endpoint {endpoint_func.__name__}. "
+                    "For simplicity, we assume a single payload argument here, "
+                    "but none was provided!"
+                )
+                picklable_payload_dict = kwargs["payload"].model_dump()
+
+                open_tasks[new_task_id] = (
+                    endpoint_func.__name__,
+                    accept,
+                    executor.submit(
+                        _run_api_function_within_worker,
+                        endpoint_func.__name__,
+                        picklable_payload_dict=picklable_payload_dict,
+                    ),
+                )
+
+                return Response(
+                    status_code=202,
+                    content=f'{{"task_id": "{new_task_id}"}}',
+                    media_type="application/json",
+                )
+            else:
+                result = endpoint_func(*args, **kwargs)
+                return create_response(result, accept)
 
         if endpoint_func.__name__ not in endpoints_to_wrap:
             return endpoint_func
         else:
[…]
