### 2025-06-06 4c6f8d1
test: test `teardown` cmd with invalid project ID
Signed-off-by: Heitor Pascoal de Bittencourt <heitorpbittencourt@gmail.com>
