### 2025-06-06 766ffee
Revert "test: test `teardown` cmd with invalid project ID"
This reverts commit 4c6f8d1d389bed6db251a6e7affba4327677f3d6.

This test does not work on MacOS and Windows. (?)

Signed-off-by: Heitor Pascoal de Bittencourt <heitorpbittencourt@gmail.com>
