### 2025-06-06 263bfa5
fix: fix teardown command crashing for wrong proj ID (#207)
#### Relevant issue or PR
Fix #208 

#### Description of changes
Handle the `NotFound` exception in the CLI.

#### Testing done
Tested manually

---------

Signed-off-by: Heitor Pascoal de Bittencourt <heitorpbittencourt@gmail.com>
