### 2025-06-06 4aba1a6
added tesseract example using differentiable QP solver qpax
