+++ examples/qp_solve/tesseract_api.py
@@ -1,0 +1,159 @@
+# Copyright 2025 Pasteur Labs. All Rights Reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+# Tesseract API module for lp_solve
+# Generated by tesseract 0.9.0 on 2025-06-04T13:11:17.208977
+import functools
+from typing import Optional
+
+import jax
+import jax.numpy as jnp
+import qpax
+from pydantic import BaseModel, Field, model_validator
+from typing_extensions import Self
+
+from tesseract_core.runtime import Array, Differentiable, Float32
+
+#
+# Schemas
+#
+
+
+class InputSchema(BaseModel):
+    Q: Differentiable[Array[(None, None), Float32]] = Field(
+        description="Quadratic cost matrix Q for the quadratic program."
+    )
+    q: Optional[Differentiable[Array[(None,), Float32]]] = Field(
+        description="Linear cost vector q for the quadratic program.",
+        default=None
+    )
+    A: Optional[Differentiable[Array[(None, None), Float32]]] = Field(
+        default=None,
+        description="Linear equality constraint matrix A for the quadratic program."
+    )
+    b: Optional[Differentiable[Array[(None,), Float32]]] = Field(
+        default=None,
+        description="Linear equality constraint rhs b for the quadratic program."
+    )
+    G: Optional[Differentiable[Array[(None, None), Float32]]] = Field(
+        description="Linear inequality constraint matrix G for the quadratic program.",
+        default=None
+    )
+    h: Optional[Differentiable[Array[(None,), Float32]]] = Field(
+        description="Linear inequality constraint rhs h for the quadratic program.",
+        default=None
+    )
+    solver_tol: Optional[Float32] = Field(
+        description="Tolerance for the solver convergence.",
+        default=1e-4
+    )
+    target_kappa: Optional[Float32] = Field(
+        description="QPAX parameter for QP relaxation.",
+        default=1e-3
+    )
+
+    @model_validator(mode="after")
+    def validate_shape_inputs(self) -> Self:
+        if self.Q.shape[0] != self.Q.shape[1]:
+            raise ValueError("Q must be a square matrix.")
+        n = self.Q.shape[0]
+        # set missing arguments to 'empty' arrays expected by qpax
+        if self.q is None:
+            self.q = jax.numpy.zeros((n,), dtype=self.Q.dtype)
+        if self.A is None:
+            self.A = jax.numpy.zeros((0, n), dtype=self.Q.dtype)
+        if self.b is None:
+            self.b = jax.numpy.zeros((0,), dtype=self.Q.dtype)
+        if self.G is None:
+            self.G = jax.numpy.zeros((0, n), dtype=self.Q.dtype)
+        if self.h is None:
+            self.h = jax.numpy.zeros((0,), dtype=self.Q.dtype)
+        # check shapes of inputs
+        if self.Q.shape[0] != self.q.shape[0]:
+            raise ValueError("Q and q must have compatible shapes.")
+        if self.A.shape[1] != self.q.shape[0]:
+            raise ValueError("A.shape[1] must be equal to number of vars.")
+        if self.A.shape[0] != self.b.shape[0]:
+            raise ValueError("A and b must have compatible shapes.")
+        if self.A.shape[0] >= self.A.shape[1]:
+            raise ValueError("A must have fewer rows than columns (m < n).")
+
+        if self.G.shape[1] != self.q.shape[0]:
+            raise ValueError("G.shape[1] must be equal to number of vars.")
+        if self.G.shape[0] != self.h.shape[0]:
+            raise ValueError("G and h must have compatible shapes.")
+        return self
+
+
+class OutputSchema(BaseModel):
[…]
