+++ examples/qp_solve/tesseract_requirements.txt
@@ -1,0 +1,10 @@
+# Tesseract requirements file
+# Generated by tesseract 0.9.0 on 2025-06-04T13:11:17.208977
+
+# Add Python requirements like this:
+# numpy==1.18.1
+qpax==0.0.9
+
+# This may contain private dependencies via SSH URLs:
+# git+ssh://git@github.com/username/repo.git@branch
+# (use `tesseract build --forward-ssh-agent` to grant the builder access to your SSH keys)
