### 2025-06-06 e2dead5
minor changes to animation
