### 2025-06-06 244ea92
Only run-tests if tests/demo/examples/core is modified
