### 2025-06-06 3f61ec2
Additional validation of service names
