### 2025-06-06 da33beb
Additional validation of service names
