### 2025-06-06 e864351
Remove bad return
