### 2025-06-09 85f0be4
chore: ✅ Update pre-commit hooks (#213)
Pre-commit hooks have been updated successfully without conflicts.

Co-authored-by: PasteurBot <hello@simulation.science>
