+++ .pre-commit-config.yaml
@@ -29,7 +29,7 @@ repos:
 
 - repo: https://github.com/astral-sh/ruff-pre-commit
   # Ruff version.
-  rev: v0.11.12
+  rev: v0.11.13
   hooks:
   # Run the linter.
   - id: ruff
