### 2025-06-09 44ae493
chore: Replace hardcoded PasteurBot email with variable (#214)
