### 2025-06-09 74807ad
chore: 📦 Update dependencies (#212)
This PR updates the lockfile to the latest versions of the dependencies.
Please review the changes and merge when ready.

---------

Co-authored-by: PasteurBot <hello@simulation.science>
Co-authored-by: Dion Häfner <dion.haefner@simulation.science>
