+++ tesseract_core/sdk/engine.py
@@ -113,6 +113,10 @@ def needs_docker(func: Callable) -> Callable:
             raise UserError(
                 "Could not reach Docker daemon, check if it is running."
             ) from ex
+        except FileNotFoundError as ex:
+            raise UserError(
+                "Docker daemon is not running, check if it is installed."
+            ) from ex
         return func(*args, **kwargs)
 
     return wrapper_needs_docker
