### 2025-06-09 8fedf4e
fix: exit with 1 when Docker executable not found
Exits with status code 1 when Docker executable is not found, instead of
crashing.

Fix #209.

Signed-off-by: Heitor Pascoal de Bittencourt <heitorpbittencourt@gmail.com>
