### 2025-06-10 #8201 @aslan-ng
> 👋 [@aslan-ng](https://github.com/aslan-ng) - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Hi @crvernon ! 👋
Thank you for considering me; I’d be happy to review the submission. Count me in!
