### 2025-06-10 2e69de0
Improve message to user
Signed-off-by: Heitor Pascoal de Bittencourt <heitorpbittencourt@gmail.com>
