### 2025-06-10 da75daa
Add handling for on push:
