+++ docs/index.md
@@ -50,6 +50,7 @@ There are several ways in which users interact with Tesseracts, for example:
 
 ## Why Tesseracts?
 
+# TESTTESTETST
 Tesseracts are primarily useful for managing **diversity**.
 
 - **Diversity of roles** *(within a team)* – The job of the software creator ends when it is packaged as a Tesseract. Users creating pipelines can focus on the high-level logic of their application, and not worry about the low-level details of how the components are implemented. Team members requesting data can inspect interfaces, docs, and schemas but don't need to dive into implementations.
