### 2025-06-10 4a2bb22
Rename error message
Co-authored-by: Alessandro Angioi <alessandro.angioi@simulation.science>
