### 2025-06-11 0b0aedc
fix: Add FileNotFoundError to docker info (#215)
#### Relevant issue or PR
#157 

#### Description of changes
Added FileNotFound exception to `needs_docker` wrapper

#### Testing done
Completely uninstalled docker and triggered the FileNotFoundError / no
docker command error.

Reinstalled package with change and now we get this error instead.

```
(tesseract) 柯 ~/work/tesseract-core » tesseract build examples/empty
 [-] Docker daemon is not running, check if it is installed.
 [x] Aborting
 ```

---------

Co-authored-by: Alessandro Angioi <alessandro.angioi@simulation.science>
