+++ .github/workflows/run_tests.yml
@@ -7,7 +7,6 @@ on:
     branches:
       - main
 
-
 concurrency:
   group: ${{ github.workflow }}-${{ github.event.pull_request.number || github.ref }}
   cancel-in-progress: true
