### 2025-06-11 0cffaac
allow skipped jobs in all-ok
