### 2025-06-11 366cadd
add skip conditions on empty matrix
