+++ .github/workflows/run_tests.yml
@@ -156,7 +156,7 @@ jobs:
             demo/**
             examples/**
             production.uv.lock
-          # .github/workflows/run_tests.yml
+            .github/workflows/run_tests.yml
 
       - name: Get available test matrix
         id: get-matrix
@@ -164,7 +164,8 @@ jobs:
           EXAMPLES_DIR: examples
           DEMO_DIR: demo
         run: |
-          if [[ "${{ steps.changed-files.outputs.any_changed }}" != "true" && ${{ github.event_name != 'push' }} ]]; then
+          # if no changed files in the current PR, skip E2E tests
+          if [[ "${{ steps.changed-files.outputs.any_changed }}" != "true" && ${{ github.event_name == 'pull_request' }} ]]; then
             echo "No changed files detected, skipping E2E tests."
             echo "matrix=[]" >> $GITHUB_OUTPUT
             echo "demos=[]" >> $GITHUB_OUTPUT
@@ -213,17 +214,17 @@ jobs:
             arch: "arm"
             docker-engine: "docker"
             unit-tesseract: "base"
+          - os: "ubuntu-24.04"
+            python-version: "3.12"
+            arch: "arm"
+            docker-engine: "docker"
+            unit-tesseract: "pyvista-arm64"
           # Run tests using Podman
           - os: "ubuntu-24.04"
             python-version: "3.12"
             arch: "x64"
             docker-engine: "podman"
             unit-tesseract: "base"
-          - os: "ubuntu-24.04"
-            python-version: "3.12"
-            arch: "arm"
-            docker-engine: "docker"
-            unit-tesseract: "pyvista-arm64"
 
       fail-fast: false
 
