+++ .github/workflows/run_tests.yml
@@ -185,7 +185,9 @@ jobs:
             find "${{ env.DEMO_DIR }}" -type d -maxdepth 1 -mindepth 1 -exec basename {} \; \
             | jq -R -s -c 'split("\n")[:-1]'
           )
+          echo "Found the following demos: $demos"
           printf 'demos=%s' "$demos" >> $GITHUB_OUTPUT
+          exit 1
 
   tests-e2e:
     needs: [get-e2e-matrix]
@@ -246,6 +248,7 @@ jobs:
         with:
           auto-update-conda: true
           python-version: ${{ matrix.python-version }}
+          channels: defaults
 
       - name: Set up uv
         uses: astral-sh/setup-uv@v6
