+++ .github/workflows/run_tests.yml
@@ -186,7 +186,6 @@ jobs:
             | jq -R -s -c 'split("\n")[:-1]'
           )
           printf 'demos=%s\n' "$demos" >> $GITHUB_OUTPUT
-          cat $GITHUB_OUTPUT
 
   tests-e2e:
     needs: [get-e2e-matrix]
