### 2025-06-11 15fcbb9
ci: Only run-tests if tests/demo/examples/core is modified (#210)
#### Relevant issue or PR
#205 

#### Description of changes
Add filter to only `run-tests` if `tests` `demo` `examples`
`tesseract-core` is modified

#### Testing done
This will be the test (if run-tests doesn't trigger)

---------

Co-authored-by: Dion Häfner <dion.haefner@simulation.science>
