### 2025-06-12 d3c4296
Expose debugpy port only in debug mode
