+++ tesseract_core/sdk/templates/docker-compose.yml
@@ -6,7 +6,9 @@ services:
     command: ["serve", "--host", "0.0.0.0", "--num-workers", "{{ service.num_workers }}"]
     ports:
       - {{ service.port }}
+    {% if service.environment.TESSERACT_DEBUG == "1" %}
       - "{{ service.debugpy_port }}:5678"
+    {% endif %}
     {%- if service.volumes %}
     volumes:
     {%- for volume in service.volumes %}
