+++ tesseract_core/sdk/docker_client.py
@@ -297,9 +297,9 @@ class Container:
         if self.attrs.get("NetworkSettings", None):
             ports = self.attrs["NetworkSettings"].get("Ports", None)
             if ports:
-                port_key = next(iter(ports))  # Get the first port key
-                if ports[port_key]:
-                    return ports[port_key][0].get("HostPort")
+                api_port_key = "8000/tcp"
+                if ports[api_port_key]:
+                    return ports[api_port_key][0].get("HostPort")
         return None
 
     @property
@@ -308,9 +308,9 @@ class Container:
         if self.attrs.get("NetworkSettings", None):
             ports = self.attrs["NetworkSettings"].get("Ports", None)
             if ports:
-                port_key = next(iter(ports))  # Get the first port key
-                if ports[port_key]:
-                    return ports[port_key][0].get("HostIp")
+                api_port_key = "8000/tcp"
+                if ports[api_port_key]:
+                    return ports[api_port_key][0].get("HostIp")
         return None
 
     @property
