### 2025-06-12 74f5615
Add debugpy port to no_compose branch
