### 2025-06-12 d1baaa4
Add instructions to where debugpy is listening to in case is up in the no compose case
