+++ tesseract_core/sdk/engine.py
@@ -600,8 +600,10 @@ def serve(
 
         if num_workers > 1:
             args.extend(["--num-workers", str(num_workers)])
-        if propagate_tracebacks:
-            args.append("--debug")
+
+        # BUG (?): verify with @dionhaefner: tesseract-runtime serve doesn't seem to have this arg
+        # if propagate_tracebacks:
+        #     args.append("--debug")
 
         # Always bind to all interfaces inside the container
         args.extend(["--host", "0.0.0.0"])
