### 2025-06-12 7c1ece2
Display ports in docker compose as well
