### 2025-06-12 e2ad3f3
Rename propagate-tracebacks -> debug
