+++ tests/sdk_tests/test_tesseract.py
@@ -117,7 +117,7 @@ def test_serve_lifecycle(mock_serving, mock_clients):
         ports=None,
         volumes=None,
         gpus=None,
-        propagate_tracebacks=True,
+        debug=True,
         num_workers=1,
         host_ip="127.0.0.1",
     )
