### 2025-06-12 15fba08
Overhaul
Signed-off-by: Heitor Pascoal de Bittencourt <heitorpbittencourt@gmail.com>
