+++ tests/sdk_tests/test_cli.py
@@ -6,6 +6,8 @@
 (Those go in endtoend_tests/test_endtoend.py.)
 """
 
+import os
+
 import pytest
 from typer.testing import CliRunner
 
@@ -35,3 +37,10 @@ def test_version(cli_runner):
     result = cli_runner.invoke(cli, ["--version"])
     assert result.exit_code == 0, result.stdout
     assert __version__ in result.stdout
+
+
+def test_bad_docker_executable_env_var(cli_runner):
+    os.environ["TESSERACT_DOCKER_EXECUTABLE"] = "not-a-docker"
+
+    result = cli_runner.invoke(cli, ["ps"])
+    assert result.exit_code == 1, result.stderr
