+++ tesseract_core/sdk/config.py
@@ -62,23 +62,20 @@ def update_config(**kwargs: Any) -> None:
     global _current_config
 
     conf_settings = {}
-    for field in RuntimeConfig.model_fields.keys():
-        env_key = f"TESSERACT_{field.upper()}"
-        if env_key in os.environ:
-            conf_settings[field] = os.environ[env_key]
+
+    if _current_config is None:
+        for field in RuntimeConfig.model_fields.keys():
+            env_key = f"TESSERACT_{field.upper()}"
+            if env_key in os.environ:
+                conf_settings[field] = os.environ[env_key]
 
     conf_settings.update(kwargs)
 
-    try:
-        config = RuntimeConfig(**conf_settings)
-    except FileNotFoundError as ex:
-        print(f"{ex}")
-        exit(1)
+    config = RuntimeConfig(**conf_settings)
     _current_config = config
 
 
 _current_config = None
-update_config()
 
 
 def get_config() -> RuntimeConfig:
