### 2025-06-12 8e11be7
Make sure config always exist
Signed-off-by: Heitor Pascoal de Bittencourt <heitorpbittencourt@gmail.com>
