+++ tesseract_core/sdk/config.py
@@ -80,4 +80,5 @@ _current_config = None
 
 def get_config() -> RuntimeConfig:
     """Return the current runtime configuration."""
+    update_config()
     return _current_config
