### 2025-06-12 760fbfe
Check string in stderr
Signed-off-by: Heitor Pascoal de Bittencourt <heitorpbittencourt@gmail.com>
