### 2025-06-12 096fe69
Remove reference to Hessian matrices
