### 2025-06-13 7f21c23
Fix wrong --debug arg
