### 2025-06-13 a79ecbd
Rename Container.debugpy_port -> Container.host_debugpy_port
