### 2025-06-13 1f95574
Fix accessing debugpy port in ports dict
