### 2025-06-13 156cc7b
Fix recycling of old name which was triggering typeguard
