### 2025-06-13 7528e1a
Remove dead code
