### 2025-06-13 2177ecc
Remove appending debugpy port to container ports
