+++ tesseract_core/sdk/cli.py
@@ -658,14 +658,6 @@ def _display_project_meta(project_id: str) -> list:
             logger.info(
                 f"Debugpy server listening at http://{host_ip}:{container.host_debugpy_port}"
             )
-            # TODO: maybe we should add an api/debugpy/... flag to these ports?
-            container_ports.append(
-                {
-                    "name": container.name,
-                    "port": container.host_debugpy_port,
-                    "ip": host_ip,
-                }
-            )
 
     return container_ports
 
