### 2025-06-13 b070cdb
Test removing none statement
