### 2025-06-13 2d25998
Minor refactoring
