### 2025-06-13 a5d86fc
Move cli initilaizaiton
