+++ tesseract_core/sdk/config.py
@@ -81,8 +81,6 @@ def update_config(**kwargs: Any) -> None:
 
 def get_config() -> RuntimeConfig:
     """Return the current runtime configuration."""
-    global _current_config
-
     if _current_config is None:
         update_config()
     return _current_config
