### 2025-06-13 6bbd5c6
Move get_config back into after logger initialization
