### 2025-06-13 11e3422
Add monkeypatch for environ
