### 2025-06-13 ca9fa51
Make sure config always exist before returning it
