+++ tesseract_core/sdk/config.py
@@ -80,4 +80,7 @@ _current_config = None
 
 def get_config() -> RuntimeConfig:
     """Return the current runtime configuration."""
+    if _current_config is None:
+        update_config()
+
     return _current_config
