### 2025-06-13 ddbebf0
feat(sdk): expose no compose in Python API
