### 2025-06-13 2e150c1
fix: Fix tests from config (#222)
- Catch and raise validation error
- Call get_config instead of update_config
- Check for none
- Use subprocess in unit test
