### 2025-06-13 0ea5b7f
Revert documentation to only mention Docker Compose
