### 2025-06-14 #8385 @aslan-ng
👋 Hi Alexander (@BoltzmannBrain) and Dion (@dionhaefner),

Great work on Tesseract Core! It tackles the challenge of reproducible, containerized pipelines with a clean CLI and built-in support for differentiable programming, all in one package. The codebase is also neat and easy to follow. It’s obvious that a lot of thought and craft went into this.

I have a few comments that I’m breaking down into three sections to make them easier to follow:
- Readme and Docs
- Code
- Paper
Please feel free to reach out if you have any questions or would like to discuss any of the points in more detail.

Thanks again for developing Tesseract-Core. It’s already a useful toolkit. I'm looking forward to seeing Tesseract Core being widely adopted!

Best,
Aslan (@aslan-ng)
