### 2025-06-14 #8385 @aslan-ng
# Readme and Docs

1. Using absolute URLs in GitHub is encouraged. Relative URLs will work fine within GitHub, but once it is pushed to PyPI, these links will become broken.
I found four instances of this issue in the `README.md`:
* `<img src="docs/static/logo-transparent.png" width="128" align="right">`
* `[Contribute](CONTRIBUTING.md)`
* `<img src="docs/img/apidoc-screenshot.png" width="600">`
* `[Apache License 2.0](LICENSE)`

2. Some elements in `README.md` run smoothly in GitHub but are not supported and won’t render properly on other platforms, such as PyPI. There are two general approaches: 1. avoiding such elements, or 2. using a different README version just for PyPI.
Here are some of those elements that I found in the README.md:
* `:package:`
* `> [!NOTE]`
* `> [!TIP]`

3. In the `README.md` file, Quick start:
“Before proceeding, make sure you have a [working installation of Docker](https://docs.docker.com/engine/install/) and a modern Python installation (Python 3.10+).”
The current instruction points users to install Docker Engine, which is Linux-centric. Consider adding a link to [Docker Desktop](https://www.docker.com/products/docker-desktop), which bundles Docker Engine, Compose, Buildx, and a GUI installer for all platforms, including Windows/macOS, so new users get everything they need in one step. You’ve already covered this in more detail under the Installation section of the [docs](https://docs.pasteurlabs.ai/projects/tesseract-core/latest/content/introduction/installation.html#basic-installation), but it may also fit neatly into the Quick Start flow.

4. In the docs, Basic installation:
“Then, verify everything is working as intended:”
```
$ tesseract list
```
On a fresh installation, if no Tesseracts have been built or registered yet, this command will display an empty table like this:

To a new user, this empty table might appear as a sign of incorrect installation. It would be helpful to explicitly mention that an empty table is expected in this case, and it indicates and confirms that the CLI is functioning correctly and that there are simply no components available yet.
