### 2025-06-14 #8385 @aslan-ng
# Code:

5. Your `test_tarball_install` invokes the CLI’s build command (in `tesseract_core/sdk/engine.py`) which calls the `engine.build_tesseract` directly, but never informs the `docker_cleanup` fixture about the newly created image. On teardown, pytest correctly complains about leftover Docker artifacts.

Two possible remedies that I can think of:
* Reuse the shared helper in common.py (the one that appends images to docker_cleanup_module["images"]) instead of calling build_docker_image directly, so cleanup is automatic.
* Explicitly append the built image tag in the CLI handler. For example, in @app.command("build") after image = engine.build_tesseract(...), do something like:
```
# pseudo-code:
docker_cleanup["images"].append(image.tags[0])
```

6. `test_serve_nonstandard_host_ip` failures:
* No-compose mode (--no-compose=True):
```
TimeoutError: Tesseract did not start in time
```
* Compose mode (--no-compose=False)
```
ConnectionError: HTTPConnectionPool(host='172.20.10.2', port=63539): Max retries exceeded
```
On inspection, the root causes appear to be when the test calls:
```
requests.get(f"http://{host_ip}:{port}/health")
```
expecting the container to bind its FastAPI server on the specified non-localhost host_ip. However, our current implementation:
* In the no-compose branch, uses:
```
ports={ "8000/tcp": (host_ip, port) }
```
which isn’t converted into a `docker run -p host_ip:port:8000` flag by our CLI wrapper.
* In the compose branch, we rely on `docker compose up --wait`, but that only guarantees the container starts.
