### 2025-06-14 #8385 @aslan-ng
# Paper

7. Since your project has a polished and professional logo, you can include it as a figure in your paper if you wish. For inspiration, see how other JOSS papers have showcased theirs, for example:

* [LightLogR: Reproducible analysis of personal light exposure data](https://joss.theoj.org/papers/10.21105/joss.07601)
* [madupite: A High-Performance Distributed Solver for Large-Scale Markov Decision Processes](https://joss.theoj.org/papers/10.21105/joss.07411)
* [STUMPY: A Powerful and Scalable Python Library for Time Series Data Mining](https://joss.theoj.org/papers/10.21105/joss.01504)
* [polyCub: An R package for Integration over Polygons](https://joss.theoj.org/papers/10.21105/joss.01056)

8. Your paper would be even stronger if you explicitly contrasted Tesseract Core with the current “status quo” tools, for instance, Snakemake, Nextflow, Kubeflow, or even plain Docker Compose. While Tesseract Core offers unique capabilities and has a competitive edge in combining such features, these other frameworks share overlapping functionality. A brief paragraph highlighting both the similarities and the key competitive advantages of Tesseract Core would give readers a clearer context about where it fits in the broader ecosystem.

9. Page 1, Statement of need:
A comma should follow “e.g.”
__“e.g. Python, Julia, C++” -> “e.g., Python, Julia, C++”__
__“e.g. cloud vs. on-prem, HPC, and supercomputers” -> e.g., “cloud vs. on-prem, HPC, and supercomputers”__

10. Page 2, Statement of need:
It is better to avoid using “+” instead of “and” in formal texts.
__“Platform + HPC engineers” -> “Platform and HPC engineers”__
or,
__“Platform + HPC engineers” -> “Platform or HPC engineers”__
or,
__“Platform + HPC engineers” -> “Platform and/or HPC engineers”__

11. Page 3, Software design:
Based on this [guide](https://learn.microsoft.com/en-us/style-guide/a-z-word-list-term-collections/c/command-line), using a hyphen is recommended when it is used as an adjective. Therefore, I recommend adding a hyphen when referring to CLI:
“command line interface (CLI)” -> “command-line interface (CLI)”

12. Page 1, Summary:
“Entry point” is the standard dictionary term for “a place or means of access.” In computer programming, an entry point is the place in a program where the execution of a program begins. ([Wikipedia](https://en.wikipedia.org/wiki/Entry_point))
“Entrypoint” is a technical compound often seen only in specific APIs or tools (e.g., Docker’s ENTRYPOINT)
In the paper, “entrypoint” seems to refer to the general term in computer programming, so I think it is better to change it like this:
__“entrypoint” -> “entry point”__

13. Page 4, Software design:
“These methods and the myriad scenarios in which they can be applied continue to be real-world validated in diverse scientific experiment, engineering physics, and industrial simulation environments.”
The sentence seems to require plural, and the neighboring words to “experiment” (“engineering physics” and “industrial simulation environments”) are also plural. So, it is better to change “experiment” to the plural form:
__“experiment” -> “experiments”__

14. Page 2, Statement of need
“Concretely, Tesseract Core addresses many of the fundamental issues that arise when building real-world SI systems; including but not limited to:”
The bit after the semicolon (“including but not limited to…”) is not a complete sentence; it’s a dependent phrase introducing a list of examples. Therefore, it should be changed to a comma:
__“real-world SI systems;” -> “real-world SI systems,”__
