### 2025-06-16 1c19d63
fix for older pydantic
