### 2025-06-16 5beaafa
feat!: Enable remote debugging (#184)
#### Relevant issue or PR
Fix #163 

#### Description of changes
  *  debugpy dependency added to the runtime.
* In debug mode, the runtime listens to port 5678 (default debugpy
port), which is forwarded to some available port on the host
* `tesseract serve` advertises debugpy listening to stuff, both in
docker-compose and no compose mode

![image](https://github.com/user-attachments/assets/128b19c4-59c0-4e27-a505-c021fad29a96)
  * Renamed `propagate-tracebacks` arg to simply `debug`
* Rather than just grabbing the one exposed port in the host_ip and
host_
[…]
