### 2025-06-16 93f45fd
Remove print
Co-authored-by: Dion Häfner <dion.haefner@simulation.science>
