### 2025-06-16 036c105
Clean up end to end test
