### 2025-06-16 27b26f9
Better error handling
Signed-off-by: Heitor Pascoal de Bittencourt <heitorpbittencourt@gmail.com>
