### 2025-06-17 #8385 @BoltzmannBrain
Thank you @aslan-ng for the great feedback, and kind comments about Tesseract!  We'll address your comments this week and notify you here when ready.
