### 2025-06-17 f3e6388
doc: new user usability improvements (#226)
Minor adjustments to README and docs, based on external user feedback.
