+++ README.md
@@ -1,4 +1,8 @@
-<img src="docs/static/logo-transparent.png" width="128" align="right">
+<picture>
+  <source media="(prefers-color-scheme: dark)" srcset="https://github.com/pasteurlabs/tesseract-core/blob/main/docs/static/logo-dark.png" width="128" align="right">
+  <img alt="" src="https://github.com/pasteurlabs/tesseract-core/blob/main/docs/static/logo-light.png" width="128" align="right">
+</picture>
+
 
 ### Tesseract Core
 
@@ -7,7 +11,7 @@ Autodiff-native, self-documenting software components for Simulation Intelligenc
 [Read the docs](https://docs.pasteurlabs.ai/projects/tesseract-core/latest/) |
 [Report an issue](https://github.com/pasteurlabs/tesseract-core/issues) |
 [Talk to the community](https://si-tesseract.discourse.group/) |
-[Contribute](CONTRIBUTING.md)
+[Contribute](https://github.com/pasteurlabs/tesseract-core/blob/main/CONTRIBUTING.md)
 
 ---
 
@@ -25,7 +29,7 @@ Tesseracts provide built-in support for [differentiable programming](https://doc
 ## Quick start
 
 > [!NOTE]
-> Before proceeding, make sure you have a [working installation of Docker](https://docs.docker.com/engine/install/) and a modern Python installation (Python 3.10+).
+> Before proceeding, make sure you have a [working installation of Docker](https://docs.docker.com/engine/install/) and a modern Python installation (Python 3.10+); if you prefer Docker Desktop for your platform, see [our extended installation instructions](https://docs.pasteurlabs.ai/projects/tesseract-core/latest/content/introduction/installation.html#basic-installation).
 
 1. Install Tesseract Core:
 
@@ -47,7 +51,7 @@ Tesseracts provide built-in support for [differentiable programming](https://doc
    ```
 
 <p align="center">
-<img src="docs/img/apidoc-screenshot.png" width="600">
+<img src="https://github.com/pasteurlabs/tesseract-core/blob/main/docs/img/apidoc-screenshot.png" width="600">
 </p>
 
 4. Run the Tesseract:
@@ -65,6 +69,6 @@ Tesseracts provide built-in support for [differentiable programming](https://doc
 
 ## License
 
-Tesseract Core is licensed under the [Apache License 2.0](LICENSE) and is free to use, modify, and distribute (under the terms of the license).
+Tesseract Core is licensed under the [Apache License 2.0](https://github.com/pasteurlabs/tesseract-core/blob/main/LICENSE) and is free to use, modify, and distribute (under the terms of the license).
 
 Tesseract is a registered trademark of Pasteur Labs, Inc. and may not be used without permission.
