### 2025-06-17 b980fd4
test(sdk.cli): don't patch env vars
Signed-off-by: Heitor Pascoal de Bittencourt <heitorpbittencourt@gmail.com>
