### 2025-06-17 3dbe55b
sdk.cli: remove dead code
