### 2025-06-17 47dcfbd
remove zip(strict=True) for py39 support
