### 2025-06-17 badedb0
fix config error handling
