### 2025-06-17 fb596bd
remove unused fixture
