### 2025-06-17 8d65152
fix: gracefully exit when Docker executable not found (#216)
#### Relevant issue or PR
Fix #209.

#### Description of changes
Exits with status code 1 when Docker executable is not found, instead of
crashing.

#### Testing done
Tested locally + extended unit tests.

---------

Signed-off-by: Heitor Pascoal de Bittencourt <heitorpbittencourt@gmail.com>
Co-authored-by: Angela Ko <angela.ko@simulation.science>
Co-authored-by: Dion Häfner <dion.haefner@simulation.science>
