+++ tesseract_core/sdk/config.py
@@ -20,13 +20,13 @@ def validate_executable(value: str | Sequence[str]) -> tuple[str, ...]:
 
     exe = shutil.which(exe)
     if exe is None:
-        raise FileNotFoundError(f"{exe} executable not found.")
+        raise ValueError(f"Executable `{value}` not found.")
 
     exe_path = Path(exe)
     if not exe_path.is_file():
-        raise FileNotFoundError(f"{exe} is not a file.")
+        raise ValueError(f"{exe} is not a file.")
     if not os.access(exe_path, os.X_OK):
-        raise PermissionError(f"{exe} is not executable.")
+        raise ValueError(f"{exe} is not executable.")
 
     return (str(exe_path.resolve()), *args)
 
@@ -67,16 +67,23 @@ def update_config(**kwargs: Any) -> None:
         if env_key in os.environ:
             conf_settings[field] = os.environ[env_key]
 
-    conf_settings.update(kwargs)
+    for field in _config_overrides:
+        conf_settings[field] = getattr(_current_config, field)
 
+    conf_settings.update(kwargs)
     config = RuntimeConfig(**conf_settings)
+
+    _config_overrides.update(set(conf_settings.keys()))
     _current_config = config
 
 
 _current_config = None
-update_config()
+_config_overrides = set()
 
 
 def get_config() -> RuntimeConfig:
     """Return the current runtime configuration."""
+    if _current_config is None:
+        update_config()
+    assert _current_config is not None
     return _current_config
