+++ tesseract_core/runtime/core.py
@@ -7,6 +7,7 @@ from pathlib import Path
 from types import ModuleType
 from typing import Any, Union
 
+import mlflow
 from pydantic import BaseModel
 
 from .config import get_config
@@ -121,7 +122,8 @@ def create_endpoints(api_module: ModuleType) -> list[Callable]:
     @assemble_docstring(api_module.apply)
     def apply(payload: ApplyInputSchema) -> ApplyOutputSchema:
         """Apply the Tesseract to the input data."""
-        out = api_module.apply(payload.inputs)
+        with mlflow.start_run():
+            out = api_module.apply(payload.inputs)
         if isinstance(out, api_module.OutputSchema):
             out = out.model_dump()
         return ApplyOutputSchema.model_validate(out)
@@ -139,7 +141,8 @@ def create_endpoints(api_module: ModuleType) -> list[Callable]:
 
             Differentiates ``jac_outputs`` with respect to ``jac_inputs``, at the point ``inputs``.
             """
-            out = api_module.jacobian(**dict(payload))
+            with mlflow.start_run():
+                out = api_module.jacobian(**dict(payload))
             return JacobianOutputSchema.model_validate(
                 out,
                 context={
@@ -162,7 +165,8 @@ def create_endpoints(api_module: ModuleType) -> list[Callable]:
             Evaluates the Jacobian vector product between the Jacobian given by ``jvp_outputs``
             with respect to ``jvp_inputs`` at the point ``inputs`` and the given tangent vector.
             """
-            out = api_module.jacobian_vector_product(**dict(payload))
+            with mlflow.start_run():
+                out = api_module.jacobian_vector_product(**dict(payload))
             return JVPOutputSchema.model_validate(
                 out, context={"output_keys": payload.jvp_outputs}
             )
@@ -181,7 +185,8 @@ def create_endpoints(api_module: ModuleType) -> list[Callable]:
             Computes the vector Jacobian product between the Jacobian given by ``vjp_outputs``
             with respect to ``vjp_inputs`` at the point ``inputs`` and the given cotangent vector.
             """
-            out = api_module.vector_jacobian_product(**dict(payload))
+            with mlflow.start_run():
+                out = api_module.vector_jacobian_product(**dict(payload))
             return VJPOutputSchema.model_validate(
                 out, context={"input_keys": payload.vjp_inputs}
             )
