+++ tesseract_core/sdk/templates/docker-compose.yml
@@ -41,3 +41,4 @@ services:
 
 networks:
   multi-tesseract-network:
+    external: true
