### 2025-06-17 2bc9000
Ruff checks
