+++ tests/sdk_tests/test_engine.py
@@ -214,7 +214,11 @@ def test_serve_tesseracts_invalid_input_args():
             engine.serve(["vectoradd"], service_names=["A", "B"])
 
         with pytest.raises(ValueError):
-            engine.serve(["vectoradd", "vectoradd"], no_compose=True, service_names=["VA1", "VA2"])
+            engine.serve(
+                ["vectoradd", "vectoradd"],
+                no_compose=True,
+                service_names=["VA1", "VA2"],
+            )
 
         with pytest.raises(ValueError):
             engine.serve(["vectoradd", "vectoradd"], service_names=["dupe", "dupe"])
@@ -224,7 +228,6 @@ def test_serve_tesseracts_invalid_input_args():
 
         with pytest.raises(ValueError):
             engine.serve(["vectoradd"], service_names=["-invalid-name"])
-
 
 
 def test_get_tesseract_images(mocked_docker):
@@ -262,7 +265,9 @@ def test_serve_tesseracts(mocked_docker):
     assert project_name_multi_tesseract
 
     # Serve and specify tesseract service names
-    project_name_multi_tesseract = engine.serve(["vectoradd", "vectoradd"], service_names=["VA1", "VA2"])
+    project_name_multi_tesseract = engine.serve(
+        ["vectoradd", "vectoradd"], service_names=["VA1", "VA2"]
+    )
     assert project_name_multi_tesseract
 
 
