+++ tesseract_core/sdk/engine.py
@@ -798,9 +798,9 @@ def _validate_service_names(service_names: list[str]) -> None:
 
     invalid_names = []
     for name in service_names:
-        if not re.match(r'^[A-Za-z0-9][A-Za-z0-9-]*$', name):
+        if not re.match(r"^[A-Za-z0-9][A-Za-z0-9-]*$", name):
             invalid_names.append(name)
-        if name[-1] == '-':
+        if name[-1] == "-":
             invalid_names.append(name)
     if invalid_names:
         raise ValueError(
