### 2025-06-18 5163c0c
Update tesseract_core/sdk/cli.py
