### 2025-06-18 5993584
feat: Add --service-names argument to `tesseract serve` so served Tesseracts can be reached by name (#206)
<!--
Please use a PR title that conforms to *conventional commits*:
"<commit_type>: Describe your change"; for example: "fix: prevent race
condition". Some other commit types are: fix, feat, ci, doc, refactor...
For a full list of commit types visit
https://www.conventionalcommits.org/en/v1.0.0/
-->

#### Relevant issue or PR
<!-- If the changes resolve an issue or follow some other PR, link to
them here. Only link something if it is directly relevant. -->

#### Description of changes
Tes
[…]
