+++ tests/endtoend_tests/test_endtoend.py
@@ -449,6 +449,41 @@ def test_tesseract_serve_with_volumes(built_image_name, tmp_path, docker_client)
             assert run_res.exit_code == 0, run_res.stderr
 
 
+def test_tesseract_serve_interop(built_image_name, docker_client, docker_cleanup):
+    cli_runner = CliRunner(mix_stderr=False)
+
+    run_res = cli_runner.invoke(
+        app,
+        [
+            "serve",
+            built_image_name,
+            built_image_name,
+            "--service-names",
+            "tess-1,tess-2",
+        ],
+        env={"COLUMNS": "1000"},
+        catch_exceptions=False,
+    )
+    assert run_res.exit_code == 0
+
+    project_meta = json.loads(run_res.stdout)
+    project_id = project_meta["project_id"]
+    docker_cleanup["project_ids"].append(project_id)
+
+    project_containers = [project_meta["containers"][i]["name"] for i in range(2)]
+
+    tess_1 = docker_client.containers.get(project_containers[0])
+
+    returncode, stdout = tess_1.exec_run(
+        [
+            "python",
+            "-c",
+            'import requests; requests.get("http://tess-2:8000/health").raise_for_status()',
+        ]
+    )
+    assert returncode == 0, stdout.decode()
+
+
 @pytest.mark.parametrize("no_compose", [True, False])
 def test_serve_nonstandard_host_ip(
     docker_client, built_image_name, docker_cleanup, free_port, no_compose
