### 2025-06-18 315b80e
clean up after test_tarball_install test
