### 2025-06-19 0a733e1
address review; add some DOIs
