+++ joss-paper/paper.bib
@@ -2,6 +2,7 @@
   title={Simulation intelligence: Towards a new generation of scientific methods},
   author={Lavin, Alexander and Krakauer, David and Zenil, Hector and Gottschlich, Justin and Mattson, Tim and Brehmer, Johann and Anandkumar, Anima and Choudry, Sanjay and Rocki, Kamil and Baydin, At{\i}l{\i}m G{\"u}ne{\c{s}} and others},
   journal={arXiv preprint arXiv:2112.03235},
+  doi={10.48550/arXiv.2112.03235},
   year={2021}
 }
 
@@ -22,6 +23,7 @@
   number={8027},
   pages={1060--1066},
   year={2024},
+  doi={10.1038/s41586-024-07744-y},
   publisher={Nature Publishing Group UK London}
 }
 
@@ -47,7 +49,8 @@
   title={Solver-in-the-loop approach to turbulence closure},
   author={Freitas, Andr{\'e} and Um, Kiwon and Desbrun, Mathieu and Buzzicotti, Michele and Biferale, Luca},
   journal={arXiv preprint arXiv:2411.13194},
-  year={2024}
+  year={2024},
+  doi={10.48550/arXiv.2411.13194},
 }
 
 @article{closure2,
@@ -58,7 +61,8 @@
   number={2},
   pages={024605},
   year={2025},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/PhysRevFluids.10.024605},
 }
 
 @software{jax2018github,
@@ -86,5 +90,25 @@
   number={POPL},
   pages={1246--1274},
   year={2023},
-  publisher={ACM New York, NY, USA}
+  publisher={ACM New York, NY, USA},
+  doi={10.1145/3571236},
+}
+
+@article{seelinger2023bridge,
+  title={UM-Bridge: Uncertainty quantification and modeling bridge},
+  author={Seelinger, Linus and Cheng-Seelinger, Vivian and Davis, Andrew and Parno, Matthew and Reinarz, Anne},
+  journal={Journal of Open Source Software},
+  volume={8},
+  number={83},
+  year={2023},
+  publisher={Open Journals},
+  doi={10.21105/joss.04748},
+}
+
+@article{brockman2016openai,
+  title={Openai gym},
+  author={Brockman, Greg and Cheung, Vicki and Pettersson, Ludwig and Schneider, Jonas and Schulman, John and Tang, Jie and Zaremba, Wojciech},
+  journal={arXiv preprint arXiv:1606.01540},
+  year={2016},
+  doi={10.48550/arXiv.1606.01540},
 }
