### 2025-06-19 fa4fbf0
fix tesseract build --forward-ssh-agent
