+++ joss-paper/paper.bib
@@ -79,7 +79,8 @@
   booktitle={2021 IEEE High Performance Extreme Computing Conference (HPEC)},
   pages={1--9},
   year={2021},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/hpec49654.2021.9622796},
 }
 
 @article{jaxautodiff,
