### 2025-06-19 #8385 @aslan-ng
Dear Dion (@dionhaefner) and Alexander (@BoltzmannBrain),

Thank you for thoughtfully addressing the feedback I provided during the review.
I’ve reviewed both the paper and the repository again, and I’m pleased to confirm that all of my comments have been fully addressed.

Best,
Aslan (@aslan-ng)
