### 2025-06-19 #8385 @aslan-ng
Dear Chris (@crvernon),

I’ve revisited both the paper and the repository, and I’m happy to confirm that all of my comments have been fully addressed. I appreciate the authors’ responsiveness and the thoughtful effort they put into refining both the paper and the repository.

I’ve also completed my reviewer checklist and am pleased to recommend this submission for publication in JOSS.

Kind regards,
Aslan (@aslan-ng)
