### 2025-06-20 8c2b1f3
feat: allow to skip checks by passing --skip-checks flag to tesseract build command
