### 2025-06-20 90d6812
Add some checks
