+++ tesseract_core/runtime/tree_transforms.py
@@ -143,6 +143,11 @@ def filter_func(
 
     def filtered_func(*args: Any) -> dict:
         if input_paths:
+            # TODO: replace this check with `zip(... strict=True)`
+            # once we stop supporting 3.9
+            assert len(input_paths) == len(args), (
+                f"Mismatch between number of given paths {len(input_paths)} and args {len(args)}."
+            )
             new_inputs = dict(zip(input_paths, args))
         else:
             if len(args) != 1:
