### 2025-06-23 1d12028
chore: ✅ Update pre-commit hooks (#236)
Pre-commit hooks have been updated successfully without conflicts.

Co-authored-by: Andrei Paleyes <apaleyes@users.noreply.github.com>
