### 2025-06-23 5cfcf7e
revert changes to sdk.cli and sdk.engine
