### 2025-06-23 ed42461
add endtoend test
