+++ examples/helloworld/tesseract_api.py
@@ -4,7 +4,7 @@
 from pydantic import BaseModel, Field
 
 
-class InputSchema(BaseModel):
+class InputSchema(Field):
     name: str = Field(description="Name of the person you want to greet.")
 
 
