### 2025-06-23 c1d509d
integrate test into existing 'test_build_generate_only' test by checking the generated Dockerfile
