### 2025-06-23 00c5d80
Add FileReference annotation for Pydantic
