### 2025-06-24 a04de1e
add extended description
