### 2025-06-27 1537b15
Add volume to docker_client and add permission tests
