+++ tests/endtoend_tests/test_docker_client.py
@@ -14,6 +14,7 @@ from common import image_exists
 
 from tesseract_core.sdk.docker_client import (
     APIError,
+    ContainerError,
     ImageNotFound,
     build_docker_image,
 )
@@ -45,6 +46,21 @@ def docker_client_built_image_name(
         except ImageNotFound:
             # already removed
             pass
+
+
+@pytest.fixture(scope="module")
+def docker_volume(docker_client):
+    # Create the Docker volume
+    volume = docker_client.volumes.create(name="docker_client_test_volume")
+    try:
+        yield volume
+    finally:
+        pass
+        # try:
+        #     volume.remove()
+        # except NotFound:
+        #     # already removed
+        #     pass
 
 
 def test_get_image(docker_client, docker_client_built_image_name, docker_py_client):
@@ -396,3 +412,69 @@ def test_compose_error(docker_client, tmp_path, docker_client_built_image_name):
         docker_client.compose.up(str(compose_file), "docker_client_compose_test")
     # Check that the container's logs were printed to stderr
     assert "Failed to start Tesseract container" in str(e.value)
+
+
+def test_volume_permissions(
+    docker_client, docker_client_built_image_name, docker_volume, docker_cleanup
+):
+    # Set up root only directory
+    def run_tesseract_with_volume(cmd: str, user: str = "root:root", volume: str = ""):
+        if not volume:
+            volume = {docker_volume.name: {"bind": "/bar", "mode": "rw"}}
+        return docker_client.containers.run(
+            docker_client_built_image_name,
+            [cmd],
+            remove=True,
+            user=user,
+            volumes=volume,
+        )
+
+    cmd = "mkdir -p /bar && echo hello > /bar/hello.txt && chmod 700 /bar"
+    _ = run_tesseract_with_volume(cmd)
+
+    # Try to read the file as UID 1000
+    read_cmd = "cat /bar/hello.txt"
+    with pytest.raises(ContainerError) as e:
+        _ = run_tesseract_with_volume(read_cmd, user="1000:1000")
+    assert "Permission denied" in str(e)
+
+    # Try to write to the folder as UID 1000
+    write_cmd = "echo hello > /bar/hello_1000.txt"
+    with pytest.raises(ContainerError) as e:
+        _ = run_tesseract_with_volume(write_cmd, user="1000:1000")
+    assert "Permission denied" in str(e)
+
+    # Grant permission to the file
+    cmd = "chmod 777 /bar"
+    _ = run_tesseract_with_volume(cmd)
+
+    # Try to read the file as UID 1000 again
+    stdout = run_tesseract_with_volume(read_cmd, user="1000:1000")
+    assert stdout == b"hello\n"
+
+    # Try to write to the folder as UID 1000 again
+    write_cmd = "echo hello > /bar/hello_1000.tx && cat /bar/hello_1000.txt"
+    stdout = run_tesseract_with_volume(write_cmd, user="1000:1000")
+    assert stdout == b"hello\n"
+
+    # Try to copy a file from a volume with permissions 700 to a volume with permission 777
+    volume_777 = docker_client.volumes.create(name="docker_client_test_volume_777")
+    docker_cleanup["volumes"].append(volume_777)
+    volume_args = {
+        docker_volume.name: {"bind": "/from", "mode": "rw"},
+        volume_777.name: {"bind": "/to", "mode": "rw"},
+    }
+    cmd = "chmod 700 /from && cp /from/hello.txt /to/hello_777.txt && cat /to/hello_777.txt"
+    stdout = run_tesseract_with_volume(cmd, volume=volume_args)
+    assert stdout == b"hello\n"
+
+    # Try to access files as UID1000
[…]
