### 2025-06-27 2369668
feat: Allow skipping checks by passing `--skip-checks` flag to the tesseract build command (#233)
Allow skipping checks by passing `--skip-checks` flag to the tesseract
build command::
```
tesseract build --skip-checks path/to/tesseract_api
```

This allows the build to proceed even in cases where Tesseract
validation or runtime checks fail (e.g. when cross-compiling or missing
mandatory files).

The flag controls two types of checks:
1. pre-build Tesseract API validation checks
2. post-build runtime checks, i.e. `tesseract check`

The runtime check behavior can be overridden by an appropriate
[…]
