+++ examples/helloworld/tesseract_api.py
@@ -14,8 +14,9 @@ class OutputSchema(BaseModel):
 
 
 def apply(inputs: InputSchema) -> OutputSchema:
-    for step in range(10):
-        metric_value = step**2  # Example metric
-        mlflow.log_metric("squared_step", metric_value, step=step)
     """Greet a person whose name is given as input."""
+    with mlflow.start_run():
+        for step in range(10):
+            metric_value = step**2  # Example metric
+            mlflow.log_metric("squared_step", metric_value, step=step)
     return OutputSchema(greeting=f"Hello {inputs.name}!")
