### 2025-06-27 000bbfa
Add mlflow example tesseract; add docs; shared volume for mlflow artifacts in docker-compose case
