+++ examples/mlflow/tesseract_requirements.txt
@@ -1,0 +1,9 @@
+# Tesseract requirements file
+# Generated by tesseract 0.9.2.dev16+g7ca45a2.d20250627 on 2025-06-27T11:44:45.333107
+
+# Add Python requirements like this:
+# numpy==1.18.1
+
+# This may contain private dependencies via SSH URLs:
+# git+ssh://git@github.com/username/repo.git@branch
+# (use `tesseract build --forward-ssh-agent` to grant the builder access to your SSH keys)
