+++ tesseract_core/sdk/templates/docker-compose-mlflow.yml
@@ -1,12 +1,37 @@
 services:
+  mlflow-init:
+    image: alpine:latest
+    user: "0:0"  # Run as root to fix volume permissions
+    command: |
+      sh -c "
+        echo 'Initializing MLflow volume permissions...'
+        mkdir -p /mlflow-data/mlruns
+        chown -R 1000:1000 /mlflow-data
+        chmod -R 755 /mlflow-data
+        echo 'MLflow volume ready for user 1000'
+      "
+    volumes:
+      - mlflow-data:/mlflow-data
+    networks:
+      - multi-tesseract-network
+
   mlflow-server:
     image: ghcr.io/mlflow/mlflow:latest
     restart: unless-stopped
-    command: mlflow server --backend-store-uri sqlite:///mlflow.db --default-artifact-root ./mlruns --host 0.0.0.0 --port 5000
+    user: 1000:1000
+    command: mlflow server --backend-store-uri sqlite:///mlflow-data/mlflow.db --default-artifact-root file:///mlflow-data/mlruns --host 0.0.0.0 --port 5000
+    volumes:
+      - mlflow-data:/mlflow-data
     networks:
       - multi-tesseract-network
     ports:
       - "5000:5000"
+    depends_on:
+      mlflow-init:
+        condition: service_completed_successfully
+
+volumes:
+  mlflow-data:
 
 # We create the network here once for all tesseracts.
 # This is to avoid creating it multiple times in subsequent docker-compose up commands.
