### 2025-06-30 6c1c013
Allow passing environment variables to tesseract serve. Important for pointing tesseract serve to MLflow servers.
