### 2025-06-30 19756ae
- fix fastapi serialization problem - fix issue with mismatched configs - add exception handling and cleanup for failed tasks
