### 2025-06-30 f2bb7da
Suppress MLflow warnings when git is not present
