+++ tesseract_core/sdk/cli.py
@@ -498,6 +498,15 @@ def serve(
             ),
         ),
     ] = None,
+    user: Annotated[
+        str | None,
+        typer.Option(
+            "--user",
+            help=(
+                "User to run the Tesseracts as e.g. '1000' or '1000:1000' (uid:gid)."
+            ),
+        ),
+    ] = None,
 ) -> None:
     """Serve one or more Tesseract images.
 
@@ -543,6 +552,7 @@ def serve(
             num_workers,
             no_compose,
             service_names_list,
+            user,
         )
     except RuntimeError as ex:
         raise UserError(
