+++ tesseract_core/sdk/templates/docker-compose.yml
@@ -4,6 +4,9 @@ services:
     image: {{ service.image }}
     restart: unless-stopped
     command: ["serve", "--host", "0.0.0.0", "--num-workers", "{{ service.num_workers }}"]
+    {%- if service.user %}
+    user: "{{ service.user }}"
+    {%- endif %}
     ports:
       - {{ service.port }}
     {% if service.environment.TESSERACT_DEBUG == "1" %}
@@ -38,6 +41,16 @@ services:
             {{ service.gpus }}
     {% endif %}
 {% endfor %}
+
+{%- if docker_volumes %}
+volumes:
+{%- for name, external in docker_volumes.items() %}
+  {{ name }}:
+    {%- if external %}
+    external: true
+    {% endif %}
+{% endfor %}
+{% endif %}
 
 networks:
   multi-tesseract-network:
