### 2025-07-01 6a85742
Add user to run cli
