### 2025-07-01 ef6a869
Fix non-iterable nonetype
