### 2025-07-01 843f913
Throw error if MLflow is not reachable (otherwise hangs by default)
