+++ tesseract_core/runtime/config.py
@@ -16,6 +16,8 @@ class RuntimeConfig(BaseModel):
     version: str = "0+unknown"
     debug: bool = False
 
+    request_timeout: float = 10.0
+
     model_config = ConfigDict(frozen=True, extra="forbid")
 
 
