### 2025-07-02 21593e9
clarifying the endpoint default and missing options in helloworld
