### 2025-07-02 #8385 @BoltzmannBrain
Hi @gartavanis, just an update: we're responding to the useful feedback you provided, namely in [this PR](https://github.com/pasteurlabs/tesseract-core/pull/243). We should have this all buttoned up for you by end-of-week :)
