### 2025-07-02 ff908ee
add tests and fix error handling
