### 2025-07-02 0af5a8b
running_in_docker fn; make dataloader example work
