+++ examples/dataloader/test-runtime.sh
@@ -8,5 +8,5 @@ cd $here
 
 export TESSERACT_API_PATH=tesseract_api.py
 
-tesseract-runtime apply '{"inputs": {"data": "@testdata/*.json"}}' --output-format json+base64 | jq
-tesseract-runtime jacobian '{"inputs": {"data": "@testdata/*.json"}, "jac_inputs": ["data.[0]"], "jac_outputs": ["data_sum"]}' | jq
+tesseract-runtime apply '{"inputs": {"data": "@testdata/*.json"}}' --output-format json+binref --output-path ./output
+# tesseract-runtime jacobian '{"inputs": {"data": "@testdata/*.json"}, "jac_inputs": ["data.[0]"], "jac_outputs": ["data_sum"]}' | jq
