### 2025-07-02 ac9f7b0
fix sdk tests
