### 2025-07-02 5f2edf8
add expected endpoints for e2e sdk tests; use Optional to pass py39 test
