### 2025-07-02 838026c
input_path example
