+++ examples/dataloader-filereference/tesseract_api.py
@@ -7,6 +7,7 @@ from typing import Annotated
 from pydantic import AfterValidator, BaseModel
 
 
+# This would go somewhere in tesseract_core.runtime
 def resolve_input_path(path: Path) -> Path:
     from tesseract_core.runtime.file_interactions import INPUT_PATH
 
@@ -17,10 +18,10 @@ def resolve_input_path(path: Path) -> Path:
 
 
 InputFileReference = Annotated[Path, AfterValidator(resolve_input_path)]
+###################################################
 
 
 class InputSchema(BaseModel):
-    # NOTE: no file references here
     data: list[InputFileReference]
 
 
@@ -30,4 +31,15 @@ class OutputSchema(BaseModel):
 
 def apply(inputs: InputSchema) -> OutputSchema:
     print(inputs.data)
+    # prints:
+    # [PosixPath('/Users/niklas/repos/pasteur/tesseract-core/examples/dataloader-filereference/testdata/sample_7.json'),
+    # PosixPath('/Users/niklas/repos/pasteur/tesseract-core/examples/dataloader-filereference/testdata/sample_6.json'),
+    # PosixPath('/Users/niklas/repos/pasteur/tesseract-core/examples/dataloader-filereference/testdata/sample_1.json'),
+    # PosixPath('/Users/niklas/repos/pasteur/tesseract-core/examples/dataloader-filereference/testdata/sample_0.json'),
+    # PosixPath('/Users/niklas/repos/pasteur/tesseract-core/examples/dataloader-filereference/testdata/sample_3.json'),
+    # PosixPath('/Users/niklas/repos/pasteur/tesseract-core/examples/dataloader-filereference/testdata/sample_2.json'),
+    # PosixPath('/Users/niklas/repos/pasteur/tesseract-core/examples/dataloader-filereference/testdata/sample_9.json'),
+    # PosixPath('/Users/niklas/repos/pasteur/tesseract-core/examples/dataloader-filereference/testdata/sample_5.json'),
+    # PosixPath('/Users/niklas/repos/pasteur/tesseract-core/examples/dataloader-filereference/testdata/sample_4.json'),
+    # PosixPath('/Users/niklas/repos/pasteur/tesseract-core/examples/dataloader-filereference/testdata/sample_8.json')]
     return OutputSchema()
