### 2025-07-02 deb0456
Add unit test for input-path Add endtoend test Change to environ variable Add input path to serve
