### 2025-07-02 8082e6d
This is a combination of 6 commits.
Add unit test for input-path

Add endtoend test

Fix path name and unit test

Change to environ variable
