+++ tesseract_core/sdk/cli.py
@@ -879,6 +879,14 @@ def run_container(
             help=("User to run the Tesseract as e.g. '1000' or '1000:1000' (uid:gid)."),
         ),
     ] = None,
+    input_path: Annotated[
+        str | None,
+        typer.Option(
+            "--input-path",
+            help="Input path to read input files from, such as local directory or S3 URI "
+            "(may be anything supported by fsspec).",
+        ),
+    ] = None,
 ) -> None:
     """Execute a command in a Tesseract.
 
@@ -940,6 +948,7 @@ def run_container(
             gpus=gpus,
             environment=environment,
             user=user,
+            input_path=input_path,
         )
 
     except ImageNotFound as e:
