### 2025-07-02 e9e4c82
Add unit test for input-path
Add endtoend test

Fix path name and unit test

Change to environ variable

Use env variable

Add input path to serve
