+++ tesseract_core/sdk/cli.py
@@ -528,6 +528,17 @@ def serve(
             ),
         ),
     ] = None,
+    input_path: Annotated[
+        str | None,
+        typer.Option(
+            "--input-path",
+            help=(
+                "Input path to read input files from, such as local directory or S3 URI "
+                "(may be anything supported by fsspec)."
+            ),
+            hidden=True,
+        ),
+    ] = None,
 ) -> None:
     """Serve one or more Tesseract images.
 
@@ -588,6 +599,7 @@ def serve(
             no_compose,
             service_names_list,
             user,
+            input_path=input_path,
         )
     except RuntimeError as ex:
         raise UserError(
@@ -883,6 +895,15 @@ def run_container(
             ),
         ),
     ] = None,
+    input_path: Annotated[
+        str | None,
+        typer.Option(
+            "--input-path",
+            help="Input path to read input files from, such as local directory or S3 URI "
+            "(may be anything supported by fsspec).",
+            hidden=True,
+        ),
+    ] = None,
 ) -> None:
     """Execute a command in a Tesseract.
 
@@ -944,6 +965,7 @@ def run_container(
             gpus=gpus,
             environment=environment,
             user=user,
+            input_path=input_path,
         )
 
     except ImageNotFound as e:
