+++ examples/dataloader-filereference/tesseract_config.yaml
@@ -1,4 +1,4 @@
-name: dataloader
+name: dataloader-filereference
 version: "1.0.0"
 description: |
   Tesseract that loads data from a folder.
