### 2025-07-02 ee30a45
WIP add environment handling to Tesseract runners
