### 2025-07-03 b478dc3
doc: Fine-tune onboarding experience (#243)
#### Relevant issue or PR

https://github.com/openjournals/joss-reviews/issues/8385#issuecomment-3003241641

#### Description of changes
- Some tweaks to docs and contributor guide.
- Add `univariate` building block to highlight an example with all
optional endpoints.

#### Testing done
:eyes:
