+++ docs/content/demo/lorenz_tesseract.md
@@ -1,5 +1,9 @@
 # Building the JAX Solver Tesseract for Lorenz-96
 
+```{seealso}
+The full Tesseract API for the Lorenz-96 model can be found in the [`demo/data-assimilation-4dvar/lorenz_tesseract`](https://github.com/pasteurlabs/tesseract-core/tree/main/demo/data-assimilation-4dvar/lorenz_tesseract) folder of the Tesseract Core repository.
+```
+
 This examples demonstrates how the JAX solver Tesseract for the Lorenz-96 model is built for the purposes of the data assimilation demo.
 
 Below is the input and output schema definition for the Lorenz Tesseract.
