### 2025-07-03 #8385 @dionhaefner
Done, please have a look @gartavanis :)

Regarding your last point: We prefer to let our future work speak for itself rather than adding forward-looking details to the paper that are bound to become stale fast as the ecosystem matures.
