### 2025-07-03 5826102
required_input_files build config, passed to RuntimeConfig through ENV
