### 2025-07-03 55b2d74
added cli check for required files and example used to test locally
