+++ examples/required_input_files/tesseract_config.yaml
@@ -1,0 +1,8 @@
+name: "required_input_files"
+version: "0.1.0"
+
+build_config:
+
+  required_input_files:
+    - parameters1.json
+    - parameters2.json
