### 2025-07-04 54a4195
raise non-timeout errors
