+++ tests/endtoend_tests/test_tesseract_sdk.py
@@ -12,17 +12,9 @@ from tesseract_core.sdk import engine
 
 expected_endpoints = {
     "apply",
-    "apply/async_start",
-    "apply/async_retrieve",
     "jacobian",
-    "jacobian/async_start",
-    "jacobian/async_retrieve",
     "jacobian_vector_product",
-    "jacobian_vector_product/async_start",
-    "jacobian_vector_product/async_retrieve",
     "vector_jacobian_product",
-    "vector_jacobian_product/async_start",
-    "vector_jacobian_product/async_retrieve",
     "health",
     "input_schema",
     "output_schema",
