### 2025-07-04 296e0c2
remove leftover print statement
Co-authored-by: Alessandro Angioi <alessandro.angioi@simulation.science>
