### 2025-07-06 ede83e4
validate inputs in main process only and validate outputs in worker process only
