### 2025-07-07 cd1bdbc
make config_override a dict
