### 2025-07-07 f1ec55d
Fix: Handle None passed as environment
