+++ tests/sdk_tests/test_cli.py
@@ -85,37 +85,36 @@ def test_config_override(
         argpairs = (
             (
                 "my-tesseract",
-                ((["name"], "my-tesseract"),),
+                {("name",): "my-tesseract"},
             ),
         )
     elif arg_to_override == "build_config.custom_build_steps":
         argpairs = (
             (
                 "[RUN foo='bar']",
-                ((["build_config", "custom_build_steps"], ["RUN foo='bar'"]),),
+                {("build_config", "custom_build_steps"): ["RUN foo='bar'"]},
             ),
             (
                 '[RUN echo "hello world"]',
-                ((["build_config", "custom_build_steps"], ['RUN echo "hello world"']),),
+                {("build_config", "custom_build_steps"): ['RUN echo "hello world"']},
             ),
         )
     elif arg_to_override == "build_config.base_image":
         argpairs = (
             (
                 "ubuntu:latest",
-                ((["build_config", "base_image"], "ubuntu:latest"),),
+                {("build_config", "base_image"): "ubuntu:latest"},
             ),
         )
     elif arg_to_override == "build_config.package_data":
         argpairs = (
             (
                 '["data/file.txt:/app/data/file.txt"]',
-                (
-                    (
-                        ["build_config", "package_data"],
-                        [("data/file.txt:/app/data/file.txt")],
-                    ),
-                ),
+                {
+                    ("build_config", "package_data"): [
+                        "data/file.txt:/app/data/file.txt"
+                    ]
+                },
             ),
         )
     else:
