### 2025-07-07 257b8c1
fix: allow to set all configs via `tesseract build --config-override` (#239)
#### Relevant issue or PR
Fixes #232 

#### Description of changes
- Parse values passed as YAML, so arguments that expect lists (like
`build_config.package_data`) can be set.
- Some more error checking and helpful error messages for good measure.
- Add more testing to ensure things work as intended.

#### Testing done
CI, old and new tests
