+++ tesseract_core/sdk/cli.py
@@ -17,6 +17,7 @@ from typing import Annotated, Any, NoReturn
 
 import click
 import typer
+import yaml
 from jinja2 import Environment, PackageLoader, StrictUndefined
 from pydantic import ValidationError as PydanticValidationError
 from rich.console import Console as RichConsole
@@ -193,27 +194,38 @@ def main_callback(
 
 def _parse_config_override(
     options: list[str] | None,
-) -> tuple[tuple[list[str], str], ...]:
+) -> dict[tuple[str, ...], Any]:
     """Parse `["path1.path2.path3=value"]` into `[(["path1", "path2", "path3"], "value")]`."""
     if options is None:
-        return ()
+        return {}
 
-    def _parse_option(option: str):
-        bad_param = typer.BadParameter(
-            f"Invalid config override {option} (must be `keypath=value`)",
-            param_hint="config_override",
-        )
-        if option.count("=") != 1:
-            raise bad_param
+    def _parse_option(option: str) -> tuple[tuple[str, ...], Any]:
+        if "=" not in option:
+            raise typer.BadParameter(
+                f'Invalid config override "{option}" (must be `keypath=value`)',
+                param_hint="config_override",
+            )
 
-        key, value = option.split("=")
-        if not key or not value:
-            raise bad_param
+        key, value = option.split("=", maxsplit=1)
+        if not re.match(r"\w[\w|\.]*", key):
+            raise typer.BadParameter(
+                f'Invalid keypath "{key}" in config override "{option}"',
+                param_hint="config_override",
+            )
 
-        path = key.split(".")
+        path = tuple(key.split("."))
+
+        try:
+            value = yaml.safe_load(value)
+        except yaml.YAMLError as e:
+            raise typer.BadParameter(
+                f'Invalid value for config override "{option}", could not parse value as YAML: {e}',
+                param_hint="config_override",
+            ) from e
+
         return path, value
 
-    return tuple(_parse_option(option) for option in options)
+    return dict(_parse_option(option) for option in options)
 
 
 @app.command("build")
