### 2025-07-07 871f377
feat: Add Volume class to docker client and --user flag to cli (#241)
#### Description of changes

- Add volume class to docker_client and docker_cleanup
- Add user flag to docker container run and serve pipelines
- Add tests to explore volume permission and user id switching

#### Testing done
Unit Testing

---------

Co-authored-by: Alessandro Angioi <alessandro.angioi@simulation.science>
