### 2025-07-07 518edb0
Ruff fix
