### 2025-07-07 b0236e6
add volume capabilities + tests to run as any user
