+++ tests/endtoend_tests/test_endtoend.py
@@ -4,7 +4,6 @@
 """End-to-end tests for Tesseract workflows."""
 
 import json
-import os
 from pathlib import Path
 
 import pytest
@@ -14,7 +13,6 @@ from common import build_tesseract, image_exists
 from typer.testing import CliRunner
 
 from tesseract_core.sdk.cli import AVAILABLE_RECIPES, app
-from tesseract_core.sdk.docker_client import ContainerError
 
 
 @pytest.fixture(scope="module")
@@ -170,6 +168,26 @@ def test_tesseract_run_stdout(built_image_name):
             print(f"failed to decode {command} stdout as JSON")
             print(run_res.stdout)
             raise
+
+
+@pytest.mark.parametrize("user", [None, "root", "1000:1000"])
+def test_run_as_user(built_image_name, user):
+    """Ensure we can run a basic Tesseract image as any user."""
+    cli_runner = CliRunner(mix_stderr=False)
+
+    run_res = cli_runner.invoke(
+        app,
+        [
+            "run",
+            built_image_name,
+            "--user",
+            user,
+            "health",
+        ],
+        catch_exceptions=False,
+    )
+    assert run_res.exit_code == 0, run_res.stderr
+    assert run_res.stdout
 
 
 @pytest.mark.parametrize("no_compose", [True, False])
@@ -399,23 +417,42 @@ def test_tesseract_serve_ports(built_image_name, port, docker_cleanup, free_port
     assert str(actual_port) in run_res.stdout
 
 
-def test_tesseract_serve_with_volumes(built_image_name, tmp_path, docker_client):
-    """Try to serve multiple Tesseracts with volume mounting."""
+@pytest.mark.parametrize("volume_type", ["bind", "named"])
+@pytest.mark.parametrize("user", [None, "root", "1000:1000"])
+def test_tesseract_serve_docker_volume(
+    built_image_name,
+    docker_client,
+    docker_volume,
+    user,
+    volume_type,
+    tmp_path,
+    docker_cleanup,
+):
+    """Test serving Tesseract with a Docker volume or bind mount.
+
+    This should cover most permissions issues that can arise with Docker volumes.
+    """
     cli_runner = CliRunner(mix_stderr=False)
     project_id = None
 
-    # Pytest creates the tmp_path fixture with drwx------ mode, we need others
-    # to be able to read and execute the path so the Docker volume is readable
-    # from within the container
-    tmp_path.chmod(0o0707)
+    dest = Path("/tesseract/output_data")
 
-    dest = Path("/foo/")
+    if volume_type == "bind":
+        # Use bind mount with a temporary directory
+        volume_to_bind = str(tmp_path)
+    elif volume_type == "named":
+        # Use docker volume instead of bind mounting
+        volume_to_bind = docker_volume.name
+    else:
+        raise ValueError(f"Unknown volume type: {volume_type}")
+
     run_res = cli_runner.invoke(
         app,
         [
             "serve",
             "--volume",
-            f"{tmp_path}:{dest}",
+            f"{volume_to_bind}:{dest}",
+            *(("--user", user) if user else []),
             built_image_name,
             built_image_name,
         ],
@@ -426,13 +463,14 @@ def test_tesseract_serve_with_volumes(built_image_name, tmp_path, docker_client)
 
     project_meta = json.loads(run_res.stdout)
     project_id = project_meta["project_id"]
+    docker_cleanup["project_ids"].append(project_id)
 
-    try:
-        tesseract0_id = project_meta["containers"][0]["name"]
-        tesseract0 = docker_client.containers.get(tesseract0_id)
-        tesseract1_id = project_meta["containers"][1]["name"]
-        tesseract1 = docker_client.containers.get(tesseract1_id)
[…]
