+++ tesseract_core/sdk/cli.py
@@ -515,7 +515,8 @@ def serve(
         typer.Option(
             "--user",
             help=(
-                "User to run the Tesseracts as e.g. '1000' or '1000:1000' (uid:gid)."
+                "User to run the Tesseracts as e.g. '1000' or '1000:1000' (uid:gid). "
+                "Defaults to the current user."
             ),
         ),
     ] = None,
@@ -844,7 +845,10 @@ def run_container(
         str | None,
         typer.Option(
             "--user",
-            help=("User to run the Tesseract as e.g. '1000' or '1000:1000' (uid:gid)."),
+            help=(
+                "User to run the Tesseract as e.g. '1000' or '1000:1000' (uid:gid). "
+                "Defaults to the current user."
+            ),
         ),
     ] = None,
 ) -> None:
