### 2025-07-07 978c11d
make maximum number of worker processes user-configurable
