+++ tesseract_core/runtime/config.py
@@ -17,6 +17,7 @@ class RuntimeConfig(BaseModel):
     debug: bool = False
 
     request_timeout: float = 10.0
+    max_num_workers: int = 4
 
     model_config = ConfigDict(frozen=True, extra="forbid")
 
