### 2025-07-07 6f81cc1
wip: limit FastAPI to 1 worker process to avoid concurrency issues.
