### 2025-07-07 5fd9a95
threadsafe set_input/output_path
