+++ tesseract_core/runtime/file_interactions.py
@@ -2,6 +2,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import json
+import threading
 import urllib.parse
 from pathlib import Path
 from typing import Any, Literal, Optional, Union, get_args
@@ -20,19 +21,32 @@ def running_in_docker() -> bool:
     return Path("/.dockerenv").exists()
 
 
-INPUT_PATH: Path = Path("/tesseract/input") if running_in_docker() else Path(".")
-OUTPUT_PATH: Path = Path("/tesseract/output") if running_in_docker() else Path(".")
+DEFAULT_INPUT_PATH: Path = Path("/tesseract/input")
+DEFAULT_OUTPUT_PATH: Path = Path("/tesseract/output")
+__PATH_CONFIG = threading.local()
+__PATH_CONFIG.input_current = DEFAULT_INPUT_PATH
+__PATH_CONFIG.output_current = DEFAULT_OUTPUT_PATH
 
 
 def set_input_path(path: PathLike) -> None:
-    global INPUT_PATH
-    INPUT_PATH = Path(path).resolve()
+    """Set the current input path."""
+    __PATH_CONFIG.input_current = Path(path).resolve()
+
+
+def get_input_path() -> Path:
+    """Get the current input path."""
+    return __PATH_CONFIG.input_current
 
 
 def set_output_path(path: PathLike) -> None:
-    global OUTPUT_PATH
-    OUTPUT_PATH = Path(path).resolve()
-    OUTPUT_PATH.mkdir(parents=True, exist_ok=True)
+    """Set the current output path."""
+    __PATH_CONFIG.output_current = Path(path).resolve()
+    # __PATH_CONFIG.output_current.mkdir(parents=True, exist_ok=True)
+
+
+def get_output_path() -> Path:
+    """Get the current output path."""
+    return __PATH_CONFIG.output_current
 
 
 def guess_format_from_path(path: PathLike) -> supported_format_type:
