+++ examples/dataloader-filereference/tesseract_api.py
@@ -5,7 +5,7 @@ import shutil
 
 from pydantic import BaseModel
 
-from tesseract_core.runtime.file_interactions import OUTPUT_PATH
+from tesseract_core.runtime.file_interactions import get_output_path
 from tesseract_core.runtime.schema_types import InputFileReference, OutputFileReference
 
 
@@ -19,8 +19,9 @@ class OutputSchema(BaseModel):
 
 def apply(inputs: InputSchema) -> OutputSchema:
     files = []
+    output_path = get_output_path()
     for source in inputs.data:
-        target = OUTPUT_PATH / source.name
+        target = output_path / source.name
         shutil.copy(source, target)
         files.append(target)
     return OutputSchema(data=files)
