### 2025-07-07 4101000
revert changes to dataloader example
