+++ tests/endtoend_tests/test_endtoend.py
@@ -480,7 +480,8 @@ def test_tesseract_serve_docker_volume(
 
         if user not in (None, "root"):
             # If we are not running as root, ensure the file is readable by the target user
-            os.chown(tmpfile, *user.split(":"))
+            uid, gid = user.split(":")
+            os.chown(tmpfile, int(uid), int(gid))
 
         exit_code, output = tesseract0.exec_run(["cat", f"{dest}/hi"])
         assert exit_code == 0
