### 2025-07-08 9381852
fix tests with no-compose
