+++ .github/workflows/run_tests.yml
@@ -274,6 +274,7 @@ jobs:
           # Podman doesn't support --cache-from
           echo "TESSERACT_DOCKER_BUILD_ARGS=" >> $GITHUB_ENV
           echo "TESSERACT_DOCKER_RUN_ARGS=--userns=keep-id" >> $GITHUB_ENV
+          echo "PODMAN_USERNS=keep-id" >> $GITHUB_ENV
 
       - name: Run test suite
         run: |
