### 2025-07-08 862435c
Offload sync endpoints to workers too; use async await to keep uvicorn responsive
