### 2025-07-09 f21789d
Add docs for passing env variables
