+++ tests/conftest.py
@@ -210,15 +210,12 @@ def docker_client():
 @pytest.fixture
 def docker_volume(docker_client):
     # Create the Docker volume
-    volume = docker_client.volumes.create(name="docker_client_test_volume")
+    volume_name = f"test_volume_{''.join(random.choices(string.ascii_lowercase + string.digits, k=8))}"
+    volume = docker_client.volumes.create(name=volume_name)
     try:
         yield volume
     finally:
-        try:
-            volume.remove()
-        except Exception:
-            # already removed
-            pass
+        volume.remove(force=True)
 
 
 @pytest.fixture(scope="module")
