### 2025-07-09 6a5bf51
Add tests for env passthrough
