### 2025-07-10 #8385 @BoltzmannBrain
Hi @gartavanis and @crvernon, let us know if additional edits (to the repo and/or paper) are needed. If none, what are the next steps? Thanks!
