### 2025-07-10 81b655c
Avoid changing dummy tesseract; consequently remove test for env passthrough on run
