+++ tests/endtoend_tests/test_endtoend.py
@@ -129,34 +129,6 @@ def test_build_generate_only(dummy_tesseract_location, skip_checks):
             assert 'RUN ["tesseract-runtime", "check"]' in docker_file_contents
 
 
-def test_env_passthrough_run(
-    docker_cleanup, docker_client, dummy_image_name, dummy_tesseract_location
-):
-    """Ensure we can pass environment variables to tesseracts when running."""
-    image_name = build_tesseract(
-        docker_client,
-        dummy_tesseract_location,
-        dummy_image_name,
-    )
-    assert image_exists(docker_client, image_name)
-    docker_cleanup["images"].append(image_name)
-
-    result = subprocess.run(
-        [
-            "tesseract",
-            "run",
-            "--env=TEST_ENV_VAR=foo",
-            dummy_image_name,
-            "apply",
-            '{"inputs": {"a": [1, 2, 3], "b": [4, 5, 6]}}',
-        ],
-        capture_output=True,
-        cwd=dummy_tesseract_location,
-    )
-    assert result.returncode == 0, result.stderr.decode()
-    assert "TEST_ENV_VAR: foo" in result.stderr.decode()
-
-
 def test_env_passthrough_serve(
     docker_cleanup, docker_client, dummy_image_name, dummy_tesseract_location
 ):
@@ -188,16 +160,13 @@ def test_env_passthrough_serve(
     project_id = project_meta["project_id"]
     tesseract_id = project_meta["containers"][0]["name"]
 
-    # Send apply request to the served Tesseract
-    res = requests.post(
-        f"http://{project_meta['containers'][0]['ip']}:{project_meta['containers'][0]['port']}/apply",
-        json={"inputs": {"a": [1, 2, 3], "b": [4, 5, 6]}},
-    )
-    assert res.status_code == 200, res.text
-
-    # Check docker logs for the environment variable
-    logs = docker_client.containers.get(tesseract_id).logs().decode("utf-8")
-    assert "TEST_ENV_VAR: foo" in logs
+    try:
+        container = docker_client.containers.get(tesseract_id)
+        exit_code, output = container.exec_run(["sh", "-c", "echo $TEST_ENV_VAR"])
+        assert exit_code == 0, f"Command failed with exit code {exit_code}"
+        assert "foo" in output.decode("utf-8"), f"Output was: {output.decode('utf-8')}"
+    except ContainerError as e:
+        pytest.fail(f"Failed to execute command in container: {e}")
 
     # Teardown the project
     run_res = cli_runner.invoke(
