### 2025-07-10 7d484d0
Remove unnecessary try block in test
