### 2025-07-10 #8385 @gartavanis
Hi @crvernon, @BoltzmannBrain, @dionhaefner,

I've reviewed the changes and I'm pleased to recommend the paper for publication in JOSS.

Best,
George
