### 2025-07-10 88c2512
Simplify env passthrough test
