### 2025-07-10 2457943
Update docs/content/using-tesseracts/advanced.md
