### 2025-07-10 a41f9e3
feat: pass env variables through `tesseract run` and `tesseract serve` (#250)
#### Relevant issue or PR

@johnbcoughlin extracted this from #229 . It's needed right away and
requires less discussion, so makes sense to merge separately.

#### Description of changes
We allow passing user-defined env variables to docker (and
docker-compose) on `tesseract run` and `tesseract serve`.

Example:

```
tesseract run --env=MLFLOW_TRACKING_URI="http://localhost:5001" helloworld apply '...'
```

Immediately relevant use case: Connecting tesseracts to user-hosted
MLflow servers.

#### Testing done
Tested m
[…]
