+++ tesseract_core/sdk/tesseract.py
@@ -26,6 +26,7 @@ class SpawnConfig:
 
     image: str
     volumes: list[str] | None
+    environment: dict[str, str] | None
     gpus: list[str] | None
     num_workers: int
     debug: bool
@@ -84,6 +85,7 @@ class Tesseract:
         image: str,
         *,
         volumes: list[str] | None = None,
+        environment: dict[str, str] | None = None,
         gpus: list[str] | None = None,
         num_workers: int = 1,
         no_compose: bool = False,
@@ -104,6 +106,7 @@ class Tesseract:
         Args:
             image: The Docker image to use.
             volumes: List of volumes to mount, e.g. ["/path/on/host:/path/in/container"].
+            environment: dictionary of environment variables to pass to the Tesseract.
             gpus: List of GPUs to use, e.g. ["0", "1"]. (default: no GPUs)
             num_workers: Number of worker processes to use. This determines how
                 many requests can be handled in parallel. Higher values
@@ -117,6 +120,7 @@ class Tesseract:
         obj._spawn_config = SpawnConfig(
             image=image,
             volumes=volumes,
+            environment=environment,
             gpus=gpus,
             num_workers=num_workers,
             debug=True,
@@ -223,6 +227,7 @@ class Tesseract:
             self._spawn_config.image,
             port=port,
             volumes=self._spawn_config.volumes,
+            environment=self._spawn_config.environment,
             gpus=self._spawn_config.gpus,
             num_workers=self._spawn_config.num_workers,
             debug=self._spawn_config.debug,
@@ -265,6 +270,7 @@ class Tesseract:
         port: str | None = None,
         host_ip: str = "127.0.0.1",
         volumes: list[str] | None = None,
+        environment: dict[str, str] | None = None,
         gpus: list[str] | None = None,
         debug: bool = False,
         num_workers: int = 1,
@@ -279,6 +285,7 @@ class Tesseract:
             [image],
             ports=ports,
             volumes=volumes,
+            environment=environment,
             gpus=gpus,
             debug=debug,
             num_workers=num_workers,
