### 2025-07-10 16d671d
Add input path to serve
