### 2025-07-10 a4f9305
Add environment to no_compose
