### 2025-07-10 c695c46
fix: Add environment to no_compose (#257)
#### Relevant issue or PR
#250 Add environment to `no_compose`

#### Description of changes
Add environment to `no_compose` and testing

#### Testing done
Unit testing
