+++ tests/endtoend_tests/test_endtoend.py
@@ -434,10 +434,7 @@ def test_tesseract_serve_ports(built_image_name, port, docker_cleanup, free_port
     assert str(actual_port) in run_res.stdout
 
 
-@pytest.mark.parametrize("use_input_path", [True, False])
-def test_tesseract_serve_with_volumes(
-    built_image_name, tmp_path, docker_client, use_input_path
-):
+def test_tesseract_serve_with_volumes(built_image_name, tmp_path, docker_client):
     """Try to serve multiple Tesseracts with volume mounting."""
     cli_runner = CliRunner(mix_stderr=False)
     project_id = None
@@ -447,19 +444,13 @@ def test_tesseract_serve_with_volumes(
     # from within the container
     tmp_path.chmod(0o0707)
 
-    volume_args = []
-    if use_input_path:
-        dest = Path("/tesseract/input_data")
-        volume_args = ["--input-path", f"{tmp_path}"]
-    else:
-        dest = Path("/foo/")
-        volume_args = ["--volume", f"{tmp_path}:{dest}"]
-
+    dest = Path("/foo/")
     run_res = cli_runner.invoke(
         app,
         [
             "serve",
-            *volume_args,
+            "--volume",
+            f"{tmp_path}:{dest}",
             built_image_name,
             built_image_name,
         ],
@@ -487,16 +478,15 @@ def test_tesseract_serve_with_volumes(
         assert exit_code == 0
         assert output.decode() == "world"
 
-        # Input path is mounted as read only volume
-        if not use_input_path:
-            # Create file inside a container and check it from the other
-            bar_file = dest / "bar"
-            exit_code, output = tesseract0.exec_run(["touch", f"{bar_file}"])
-            assert exit_code == 0
-            exit_code, output = tesseract1.exec_run(["cat", f"{bar_file}"])
-            assert exit_code == 0
-            # The file should exist outside the container
-            assert (tmp_path / "bar").exists()
+        # Create file inside a container and check it from the other
+        bar_file = dest / "bar"
+        exit_code, output = tesseract0.exec_run(["touch", f"{bar_file}"])
+        assert exit_code == 0
+        exit_code, output = tesseract1.exec_run(["cat", f"{bar_file}"])
+        assert exit_code == 0
+
+        # The file should exist outside the container
+        assert (tmp_path / "bar").exists()
     finally:
         if project_id:
             run_res = cli_runner.invoke(
@@ -668,20 +658,13 @@ def test_serve_nonstandard_host_ip(
         requests.get(f"http://localhost:{project_container.host_port}/health")
 
 
-@pytest.mark.parametrize("use_input_path", [True, False])
-def test_tesseract_cli_options_parsing(built_image_name, tmpdir, use_input_path):
+def test_tesseract_cli_options_parsing(built_image_name, tmpdir):
     cli_runner = CliRunner(mix_stderr=False)
 
     tmpdir.chmod(0o0707)
 
     examples_dir = Path(__file__).parent.parent.parent / "examples"
     example_inputs = examples_dir / "vectoradd" / "example_inputs.json"
-
-    if use_input_path:
-        additional_options = ["--input-path", str(examples_dir)]
-        example_inputs = "vectoradd/example_inputs.json"
-    else:
-        additional_options = []
 
     test_commands = (
         ["apply", "-f", "json+binref", "-o", str(tmpdir), f"@{example_inputs}"],
@@ -695,7 +678,6 @@ def test_tesseract_cli_options_parsing(built_image_name, tmpdir, use_input_path)
             [
                 "run",
                 built_image_name,
-                *additional_options,
                 *args,
             ],
             catch_exceptions=False,
