+++ tesseract_core/runtime/cli.py
@@ -82,24 +82,12 @@ def _parse_arg_callback(
         return value, base_dir
 
     if value.startswith("@"):
-        value = value[1:]
-        tesseract_input_path = os.environ.get("TESSERACT_INPUT_PATH")
-        if tesseract_input_path:
-            if not os.access(tesseract_input_path, os.R_OK):
-                raise click.BadParameter(
-                    f"Tesseract input path {tesseract_input_path} is not accessible."
-                )
-            value = os.path.join(tesseract_input_path, value)
-            typer.echo(
-                f"Resolving @path to {value}. \n"
-                "Clear the TESSERACT_INPUT_PATH environment variable to disable this behavior."
-            )
-        base_dir = Path(value).parent
-        value_format = guess_format_from_path(value)
+        base_dir = Path(value[1:]).parent
+        value_format = guess_format_from_path(value[1:])
         try:
-            value_bytes = read_from_path(value)
+            value_bytes = read_from_path(value[1:])
         except Exception as e:
-            raise click.BadParameter(f"Could not read data from path {value}.") from e
+            raise click.BadParameter(f"Could not read data from path {value}") from e
     else:
         # Data given directly via the CLI is always in JSON format
         value_format = "json"
@@ -494,7 +482,6 @@ def main() -> None:
                 file=sys.stderr,
             )
             sys.exit(1)
-
         cli = _add_user_commands_to_cli(tesseract_runtime, out_stream=orig_stdout)
         cli(auto_envvar_prefix="TESSERACT_RUNTIME")
 
