### 2025-07-11 cd8d6ed
Filereference schemas
rename example

prevent leaking sensitive info?

refactor to use env variables correctly? still need fixes for SDK CLI
