### 2025-07-11 84214a0
cli/sdk fixes
