### 2025-07-11 19d742a
docstring
