### 2025-07-11 11177e7
feat: allow to run T containers as any user, for better volume permission handling (#253)
#### Relevant issue or PR
n/a

#### Description of changes
This enables users to execute Tesseracts as any system user, even in the
presence of volume mounts.

We achieve this in the following ways:
1. Default to running Tesseracts with the same uid as the current host
user. This fixes most permission issues right away (since everything
that's readable from the outside also becomes readable from the inside).
2. Create mount points ahead of time and ensure they have lax
permissions (rw for everyone). This 
[…]
