### 2025-07-11 171b9bf
chore: 🚢 release v0.10.0 (#242)
This PR contains the generated changelog for the release v0.10.0.

⚠️ **Merging this PR will immediately trigger a new release**. ⚠️

To specify additional release notes, please edit this comment after the
following line.

---

This is a big release with **several new features and quality-of-life
improvements**:

- :tada: Tesseracts now support remote debugging via
[`debugpy`](https://github.com/microsoft/debugpy). For example, this
allows users to connect to a running Tesseract via VSCode for
interactive debugging sessions, even in cases where the Tesseract is

[…]
