### 2025-07-11 a3d75b0
refactor to use env variables correctly? still need fixes for SDK CLI
