+++ tesseract_core/runtime/schema_types.py
@@ -1,6 +1,7 @@
 # Copyright 2025 Pasteur Labs. All Rights Reserved.
 # SPDX-License-Identifier: Apache-2.0
 
+import os
 from abc import ABCMeta
 from enum import IntEnum
 from functools import partial
@@ -389,12 +390,23 @@ def _resolve_input_path(path: Path) -> Path:
 
 
 def _strip_output_path(path: Path) -> Path:
-    from tesseract_core.runtime.file_interactions import get_output_path
+    from tesseract_core.runtime.file_interactions import running_in_docker
 
-    output_path = get_output_path()
+    client_output_path = os.environ.get("TESSERACT_CLIENT_OUTPUT_PATH", None)
+    if client_output_path is None:
+        raise ValueError("Output path not set")
+    else:
+        client_output_path = Path(client_output_path)
+
+    if running_in_docker():
+        output_path = Path("/tesseract/output_path")
+    else:
+        output_path = client_output_path
+
     if path.is_relative_to(output_path):
         return path.relative_to(output_path)
-    return path
+    else:
+        return path
 
 
 InputFileReference = Annotated[Path, AfterValidator(_resolve_input_path)]
