+++ tesseract_core/runtime/file_interactions.py
@@ -15,31 +15,47 @@ PathLike = Union[str, Path]
 supported_format_type = Literal["json", "msgpack", "json+base64", "json+binref"]
 SUPPORTED_FORMATS = get_args(supported_format_type)
 
+os.environ["TESSERACT_CONTAINER_INPUT_PATH"] = "/tesseract/input_path"
+os.environ["TESSERACT_CONTAINER_OUTPUT_PATH"] = "/tesseract/output_path"
+
 
 def running_in_docker() -> bool:
     """Check if tesseract-runtime is running inside a Docker container."""
     return Path("/.dockerenv").exists()
 
 
-def set_input_path(path: PathLike) -> None:
-    """Set the current input path."""
-    os.environ["TESSERACT_INPUT_PATH"] = str(path)
-
-
 def get_input_path() -> Path:
     """Get the current input path."""
-    return Path(os.environ.get("TESSERACT_INPUT_PATH", "/tesseract/input_path"))
-
-
-def set_output_path(path: PathLike) -> None:
-    """Set the current output path."""
-    Path(path).mkdir(parents=True, exist_ok=True)
-    os.environ["TESSERACT_OUTPUT_PATH"] = str(path)
+    path = os.environ.get("TESSERACT_CLIENT_INPUT_PATH", None)
+    if path is None:
+        raise ValueError("Input path not set.")
+    if running_in_docker():
+        return Path("/tesseract/input_path")
+    return Path(path)
 
 
 def get_output_path() -> Path:
     """Get the current output path."""
-    return Path(os.environ.get("TESSERACT_OUTPUT_PATH", "/tesseract/output_path"))
+    path = os.environ.get("TESSERACT_CLIENT_OUTPUT_PATH", None)
+    if path is None:
+        raise ValueError("Output path not set.")
+    if running_in_docker():
+        return Path("/tesseract/output_path")
+    return Path(path)
+
+
+def set_client_input_path(path) -> None:
+    path = Path(path).resolve()
+    os.environ["TESSERACT_CLIENT_INPUT_PATH"] = str(path)
+    if not running_in_docker():
+        path.mkdir(parents=True, exist_ok=True)
+
+
+def set_client_output_path(path) -> None:
+    path = Path(path).resolve()
+    os.environ["TESSERACT_CLIENT_OUTPUT_PATH"] = str(path)
+    if not running_in_docker():
+        path.mkdir(parents=True, exist_ok=True)
 
 
 def guess_format_from_path(path: PathLike) -> supported_format_type:
