+++ tesseract_core/sdk/tesseract.py
@@ -17,6 +17,11 @@ import requests
 from pydantic import BaseModel, TypeAdapter, ValidationError
 from pydantic_core import InitErrorDetails
 
+from tesseract_core.runtime.file_interactions import (
+    set_client_input_path,
+    set_client_output_path,
+)
+
 from . import engine
 
 PathLike = str | Path
@@ -64,7 +69,6 @@ class Tesseract:
         self._serve_context = None
         self._lastlog = None
         self._client = HTTPClient(url)
-        self._path_config = {"input_current": None, "output_current": None}
 
     @classmethod
     def from_url(cls, url: str) -> Tesseract:
@@ -127,11 +131,18 @@ class Tesseract:
         """
         obj = cls.__new__(cls)
 
+        if environment is None:
+            environment = {}
+
         volumes = []
         if input_path is not None:
-            volumes.append(f"{Path(input_path).resolve()}:/tesseract/input_path:ro")
+            input_path = Path(input_path).resolve()
+            environment["TESSERACT_CLIENT_INPUT_PATH"] = str(input_path)
+            volumes.append(f"{input_path}:/tesseract/input_path:ro")
         if output_path is not None:
-            volumes.append(f"{Path(output_path).resolve()}:/tesseract/output_path:rw")
+            output_path = Path(output_path).resolve()
+            environment["TESSERACT_CLIENT_OUTPUT_PATH"] = str(output_path)
+            volumes.append(f"{output_path}:/tesseract/output_path:rw")
 
         obj._spawn_config = SpawnConfig(
             image=image,
@@ -145,10 +156,6 @@ class Tesseract:
         obj._serve_context = None
         obj._lastlog = None
         obj._client = None
-        obj._path_config = {
-            "input_current": Path("/tesseract/input_path"),
-            "output_current": Path("/tesseract/output_path"),
-        }
         return obj
 
     @classmethod
@@ -192,19 +199,16 @@ class Tesseract:
                     f"Cannot load Tesseract API from {tesseract_api_path}"
                 ) from ex
 
+        if input_path is not None:
+            set_client_input_path(input_path)
+        if output_path is not None:
+            set_client_output_path(output_path)
+
         obj = cls.__new__(cls)
         obj._spawn_config = None
         obj._serve_context = None
         obj._lastlog = None
         obj._client = LocalClient(tesseract_api)
-        obj._path_config = {
-            "input_current": Path(input_path).resolve()
-            if input_path is not None
-            else Path.cwd(),
-            "output_current": Path(output_path).resolve()
-            if output_path is not None
-            else Path.cwd(),
-        }
         return obj
 
     def __enter__(self) -> Tesseract:
@@ -214,18 +218,6 @@ class Tesseract:
         """
         if self._serve_context is not None:
             raise RuntimeError("Cannot serve the same Tesseract multiple times.")
-
-        if self._path_config["input_current"] is not None:
-            from tesseract_core.runtime.file_interactions import set_input_path
-
-            set_input_path(self._path_config["input_current"])
-
-        if self._path_config["output_current"] is not None:
-            from tesseract_core.runtime.file_interactions import set_output_path
-
-            set_output_path(self._path_config["output_current"])
-            if self._spawn_config is None:
-                self._path_config["output_current"].mkdir(parents=True, exist_ok=True)
 
         if self._client is not None:
             # Tesseract is already being served -> no-op
