### 2025-07-11 4335b9c
Combined various env variables into one
