+++ tesseract_core/runtime/cli.py
@@ -28,8 +28,8 @@ from tesseract_core.runtime.file_interactions import (
     load_bytes,
     output_to_bytes,
     read_from_path,
-    set_client_input_path,
-    set_client_output_path,
+    set_input_path,
+    set_output_path,
     write_to_path,
 )
 from tesseract_core.runtime.finite_differences import (
@@ -380,7 +380,7 @@ def _create_user_defined_cli_command(
         **optional_args: Any,
     ):
         if input_path:
-            set_client_input_path(input_path)
+            set_input_path(input_path)
 
         if output_format == "json+binref" and output_path is None:
             raise ValueError("--output-path must be specified for json+binref format")
@@ -403,7 +403,7 @@ def _create_user_defined_cli_command(
                 ) from e
 
         if output_path:
-            set_client_output_path(output_path)
+            set_output_path(output_path)
 
         result = user_function(**user_function_args)
         result = output_to_bytes(result, output_format, output_path)
