### 2025-07-11 af851c1
Fix unit tests
