### 2025-07-11 f10d824
Add filereference examples test
