### 2025-07-11 62dfbbb
Add dataloader-filereference examples test
