### 2025-07-11 405a4bb
Fix config for test_examples
