### 2025-07-14 5a2945c
complete merge
