### 2025-07-14 0696657
@ syntax is not related to input_path
