### 2025-07-22 #8385 @crvernon
@dionhaefner can you sync your `joss` branch (the one I am assuming is most up to date for the paper mods) with your `main` branch that has the most recent changes?  Then you can release / mint your `joss` branch reflecting all current modifications resulting from the review and what is described in the paper.
